Address review: split sync.rs, fix pull-limit const naming/docs, revert page size to 100
Six fixes from review on the sync pagination PR:

- Split the 2938-line crates/spelunk-cli/.../memory/sync.rs into a sync/
  directory module by concern: mod.rs (command wiring), push/ (push_local +
  chunking/vector/counting tests, itself split further), pull.rs (pull +
  pagination), round.rs (sync_round), test_support.rs (shared fixtures).
- Moved the misplaced doc comment off `pull_since` and onto the const it
  was actually describing.
- Renamed `PULL_PAGE_LIMIT` to `MEMORY_SINCE_MAX_LIMIT`: it's scoped to the
  `/memory/since` pull path only, not a generic page-size limit.
- Rewrote that const's doc to describe only the server's maximum accepted
  `limit`, not what this client sends.
- Added a separate `MEMORY_SINCE_PULL_LIMIT` const for the value actually
  requested, and pointed the one remaining magic-number test assertion at
  it instead of a literal.
- Reverted the requested page size from 500 back to 100 (the server's own
  default): the pagination loop, not the per-page size, is what fixes the
  truncation bug, so there's no correctness reason to request the max, and
  100 is kinder to slow connections and large entries. Retuned the
  pagination test fixtures accordingly (100-entry "full" pages instead of
  500), keeping each test's shape unchanged.

Full workspace test suite green, clippy clean (default and
--no-default-features), fmt clean.
