--- crates/spelunk-cli/src/cli/cmd/memory/sync.rs
@@ -1,2938 +1,0 @@
-//! `spelunk sync` and `spelunk memory pull` — two-way local↔cloud memory sync.
-//!
-//! - `pull`: delta-pull from `GET /memory/since?since_id=<cursor>` and apply
-//!   locally. The cursor is the max cloud `remote_id` already synced
-//!   (`MAX(remote_id)` over local notes), a UUIDv7 — not a wall-clock watermark
-//!   (decision #183), so it is immune to local↔remote clock drift.
-//! - `sync`: push local rows `WHERE remote_id IS NULL` via `POST /memory/batch`
-//!   (batched, not N single POSTs), then pull everything after the cursor and
-//!   apply both.
-//!
-//! Properties:
-//! - **Idempotent.** Identity is the stable UUID; pushes carry it as the cloud
-//!   `external_id` and pulls record the cloud UUID as the local `remote_id` and
-//!   dedupe on it, so re-running never duplicates. Same-millisecond boundary
-//!   entries are harmless: the cursor comparison is strict (`>`) and pull
-//!   dedupes by `remote_id`, so a re-applied boundary row is a no-op.
-//! - **Entity-id reconciled.** A pulled entry that matches an existing local
-//!   row's `kind`/`title`/`body` (`entity_id`) reuses that row instead of
-//!   adding a duplicate: the row adopts the pulled `remote_id` if it had
-//!   none, and archival propagates the same never-un-archive way a matching
-//!   `remote_id` does; semantic-dup detection is the server's job (it flags
-//!   `contradicts`).
-//! - **Lifecycle propagation.** `supersedes` and archive/tombstone state travel
-//!   in both directions (previously hard-coded `None`/dropped).
-//! - **Text-only by default; optional pushed vector.** A push ships no vector
-//!   and the server backfills the embedding (embedding-model conformance) —
-//!   unless the server advertises `accepts_pushed_vectors`, in which case an
-//!   entry with a local fp32/896 embedding carries it so the server stores it
-//!   as-is instead of re-embedding.
-
-use anyhow::{Context, Result};
-
-use super::{MemoryPullArgs, MemorySyncArgs};
-use crate::{
-    capability,
-    cli::cmd::auth_api,
-    config::Config,
-    storage::{BatchPushItem, CloudSyncClient, MemoryStore},
-};
-
-/// How many entries go in each `POST /memory/batch` request. Kept small so even
-/// the worst case (text-only, the server re-embedding a full chunk on a cold
-/// embedder) finishes with wide margin under the request timeout, and a
-/// with-vectors chunk (N x 896 fp32) is a sub-megabyte JSON body rather than
-/// multi-MB; still large enough that a few hundred entries is only a handful of
-/// requests. A finer chunk also bounds the loss on a mid-push failure to at most
-/// this many entries before the resumable re-run continues. Fixed, not
-/// configurable: no evidence a user needs to tune it, and a constant is trivial
-/// to adjust later.
-const PUSH_BATCH_CHUNK_SIZE: usize = 50;
-
-/// Resolve the project slug to sync into, or halt with actionable guidance.
-///
-/// Precedence: an explicit `--project <slug>` overrides any configured
-/// `project_id`; otherwise the configured `project_id` is used. When neither is
-/// present the call **halts** — sync never auto-derives a name from the folder
-/// or git remote (founder decision 2026-07-01, project-taxonomy). The returned
-/// slug is sent verbatim to the server, which lazily creates the project on
-/// first sync and reuses it on subsequent syncs.
-fn resolve_sync_project(cli_project: Option<&str>, cfg: &Config) -> Result<String> {
-    cli_project
-        .map(str::to_string)
-        .or_else(|| cfg.project_id.clone())
-        .filter(|s| !s.trim().is_empty())
-        .ok_or_else(|| {
-            anyhow::anyhow!(
-                "No project specified. Re-run as `spelunk sync --project <slug>` \
-                 to choose the cloud project to sync into.\n\
-                 (The project is created on first sync from the slug you pass; \
-                 the slug is never guessed from the folder or git remote.)"
-            )
-        })
-}
-
-/// Resolve the cloud sync target (base URL, server-side project id, key).
-///
-/// Sync always speaks to an explicit `server_url`, not the inference loopback.
-/// Errors with actionable guidance when the server is missing (via
-/// [`capability::require_explicit_server_url`], the same guard `memory push` uses),
-/// or when no project slug is available (see [`resolve_sync_project`]).
-///
-/// `feature` names the calling command (`"sync"` or `"memory pull"`) for the
-/// error message. `cli_project` is the optional `--project <slug>` override;
-/// when `None` the configured `project_id` is used.
-///
-/// The bearer key is resolved through [`auth_api::ensure_fresh_server_key`] so a
-/// WorkOS access token that has expired since `spelunk login` is refreshed (and
-/// the rotated tokens persisted) before the cloud-api call, rather than 401-ing.
-async fn sync_target(
-    feature: &str,
-    cfg: &Config,
-    cli_project: Option<&str>,
-) -> Result<(String, String, Option<String>)> {
-    let base_url = capability::require_explicit_server_url(feature, cfg)?;
-    let project_id = resolve_sync_project(cli_project, cfg)?;
-    let key = auth_api::ensure_fresh_server_key(cfg, &base_url).await?;
-    Ok((base_url, project_id, key))
-}
-
-/// `spelunk memory pull` — one-way delta pull + apply.
-pub async fn memory_pull(
-    _args: MemoryPullArgs,
-    mem_path: &std::path::Path,
-    cfg: &Config,
-) -> Result<()> {
-    let tier = capability::get_tier(cfg).await;
-    capability::require_tier1("memory pull", tier, cfg.server_url.as_deref())?;
-    let (base_url, project_id, key) = sync_target("memory pull", cfg, None).await?;
-
-    let local = MemoryStore::open(mem_path)
-        .with_context(|| format!("opening local memory at {}", mem_path.display()))?;
-    let client = CloudSyncClient::new(
-        &base_url,
-        &project_id,
-        key.as_deref(),
-        cfg.server_ca.as_deref().map(std::path::Path::new),
-    )?;
-
-    let pulled = pull_and_apply(&local, &client).await?;
-    println!("Pull complete. Applied {pulled} new remote entries.");
-    Ok(())
-}
-
-/// `spelunk sync` (and `spelunk memory push`'s successor) — two-way sync.
-pub async fn memory_sync(
-    args: MemorySyncArgs,
-    mem_path: &std::path::Path,
-    cfg: &Config,
-) -> Result<()> {
-    let tier = capability::get_tier(cfg).await;
-    capability::require_tier1("sync", tier, cfg.server_url.as_deref())?;
-    let (base_url, project_id, key) = sync_target("sync", cfg, args.project.as_deref()).await?;
-
-    let src_path = args.source.as_deref().unwrap_or(mem_path);
-    let local = MemoryStore::open(src_path)
-        .with_context(|| format!("opening local memory at {}", src_path.display()))?;
-    let client = CloudSyncClient::new(
-        &base_url,
-        &project_id,
-        key.as_deref(),
-        cfg.server_ca.as_deref().map(std::path::Path::new),
-    )?;
-
-    // ── Two-phase reconciliation: pull, push, pull again off the same
-    // pre-round cursor. See `sync_round` for why a plain push-then-pull (or
-    // even pull-then-push) reorder is not sufficient. ───────────────────────
-    let accepts_pushed_vectors = tier.caps().is_some_and(|c| c.accepts_pushed_vectors);
-    let SyncRoundOutcome { pushed, pulled } = sync_round(
-        &local,
-        &client,
-        args.include_archived,
-        accepts_pushed_vectors,
-    )
-    .await?;
-
-    if pushed.attempted == 0 {
-        println!(
-            "Nothing to push — {} entries already synced. Applied {} new remote entries.",
-            pushed.already_synced, pulled
-        );
-    } else if let Some(reason) = pushed.interrupted.as_deref() {
-        // A chunk failed mid-push. The pull half already ran (it is independently
-        // useful, so there is no reason to withhold remote changes), but the
-        // command must fail loud and tell the user how to resume: the chunks that
-        // landed are durably stamped, so a re-run pushes only the remainder and
-        // already-pushed entries come back as 207 `skipped`.
-        anyhow::bail!(
-            "Pushed {} of {} entries, then stopped: {reason}. \
-             Re-run to resume (already-pushed entries are skipped). \
-             Pull applied {} new remote entries.",
-            pushed.created + pushed.skipped,
-            pushed.attempted,
-            pulled
-        );
-    } else if pushed.created == 0 && pushed.skipped == 0 {
-        // Total push failure: nothing durably landed cloud-side. The pull
-        // already ran (it's an independent, unconditionally useful step —
-        // there's no reason to withhold remote changes just because the push
-        // half failed), but the command must still fail loud: a caller
-        // skimming for "Sync complete" or checking only the exit code must
-        // not read this as success.
-        anyhow::bail!(
-            "Sync failed: 0 of {} push entries reached the server ({} failed); \
-             pull still applied {} new remote entries.",
-            pushed.attempted,
-            pushed.failed,
-            pulled
-        );
-    } else if pushed.failed > 0 {
-        println!(
-            "Sync complete. Pushed {} entries (created {}, skipped {}, {} failed), applied {} new remote entries.",
-            pushed.attempted, pushed.created, pushed.skipped, pushed.failed, pulled
-        );
-    } else {
-        println!(
-            "Sync complete. Pushed {} entries (created {}, skipped {}), applied {} new remote entries.",
-            pushed.attempted, pushed.created, pushed.skipped, pulled
-        );
-    }
-    Ok(())
-}
-
-/// Outcome of a push pass (shared by `sync` and the one-way `memory push`).
-#[derive(Debug)]
-pub(super) struct PushSummary {
-    /// Rows actually sent to `push_batch` (the `live` set) — not the raw
-    /// pre-filter row count, which would over-report when rows are already
-    /// synced (`remote_id` already set) and no request is made at all.
-    pub attempted: usize,
-    /// Tallied from `results[].status`, not the server's own aggregate
-    /// `created`/`skipped` ints — the two are independent wire fields
-    /// (`BatchPushResult`) and can diverge (a server has been observed
-    /// reporting aggregate `created: 0` for a batch whose per-item results
-    /// showed entries durably persisted). `results[]` is the reconciled
-    /// signal.
-    pub created: u32,
-    pub skipped: u32,
-    /// Items whose status did not affirmatively mean "durably persisted"
-    /// — anything other than `created`/`skipped` (`failed`, or an
-    /// unrecognized status riding along with a result). Kept separate so a
-    /// partial-failure batch still reports its real successes instead of
-    /// reading as "nothing happened".
-    pub failed: u32,
-    /// Non-archived rows already carrying a `remote_id` — i.e. previously
-    /// synced and excluded from `attempted`. Lets callers report an honest
-    /// "nothing to push" message instead of implying a push happened.
-    pub already_synced: usize,
-    /// `Some(reason)` when a chunk failed mid-push and the loop stopped before
-    /// the remaining chunks (see [`push_local`]); `None` on a clean run. Lets the
-    /// command layer report honest partial progress plus a resume hint and exit
-    /// non-zero, instead of discarding the chunks that already landed.
-    pub interrupted: Option<String>,
-}
-
-/// One-way push entry point reused by `spelunk memory push`.
-///
-/// `accepts_pushed_vectors` mirrors the destination server's `/v1/health`
-/// capability: when true, each entry that has a local embedding
-/// carries its fp32/896 vector so the server stores it as-is; when false the
-/// push is text-only and the server re-embeds.
-pub(super) async fn push_local_oneway(
-    local: &MemoryStore,
-    client: &CloudSyncClient,
-    include_archived: bool,
-    accepts_pushed_vectors: bool,
-) -> Result<PushSummary> {
-    push_local(local, client, include_archived, accepts_pushed_vectors).await
-}
-
-/// Push local entries to the cloud in batches, then propagate tombstones for any
-/// archived rows that exist cloud-side. Each entry is text-only unless
-/// `accepts_pushed_vectors` is set and the row has a local fp32/896 embedding,
-/// in which case that vector is attached (the server stores it without
-/// re-embedding).
-///
-/// On a chunk-level failure the push stops at the first failed chunk and returns
-/// a summary marked `interrupted` (rather than `?`-propagating and discarding
-/// the chunks that already landed); the already-stamped chunks make the next run
-/// resume from the remainder. See [`push_local_reporting`].
-async fn push_local(
-    local: &MemoryStore,
-    client: &CloudSyncClient,
-    include_archived: bool,
-    accepts_pushed_vectors: bool,
-) -> Result<PushSummary> {
-    push_local_reporting(
-        local,
-        client,
-        include_archived,
-        accepts_pushed_vectors,
-        |done, total| {
-            // Transient cumulative status on stderr; the final one-line summary
-            // stays on stdout in the command layer.
-            eprintln!("Pushed {done}/{total}…");
-        },
-    )
-    .await
-}
-
-/// [`push_local`] with the per-chunk progress emission injected as `on_progress`
-/// (called with cumulative `done`/`total` after each chunk that lands), so tests
-/// can observe the progress sequence without capturing stderr. `push_local`
-/// passes the stderr-writing closure.
-async fn push_local_reporting(
-    local: &MemoryStore,
-    client: &CloudSyncClient,
-    include_archived: bool,
-    accepts_pushed_vectors: bool,
-    mut on_progress: impl FnMut(usize, usize),
-) -> Result<PushSummary> {
-    let rows = local.rows_for_sync(include_archived)?;
-    if rows.is_empty() {
-        return Ok(PushSummary {
-            attempted: 0,
-            created: 0,
-            skipped: 0,
-            failed: 0,
-            already_synced: 0,
-            interrupted: None,
-        });
-    }
-
-    // Split into live entries (batch-created/upserted by external_id) and
-    // archived entries already known to the cloud (tombstoned via DELETE).
-    let mut created = 0u32;
-    let mut skipped = 0u32;
-    let mut failed = 0u32;
-    // Set once a chunk fails: the loop stops and the tombstone pass is skipped.
-    let mut interrupted: Option<String> = None;
-
-    // Push set (decision #183): live entries not yet on the cloud — i.e.
-    // `WHERE remote_id IS NULL`. Already-synced rows carry a `remote_id` and are
-    // skipped here (the cloud already has them; re-pushing would only earn a 207
-    // `skipped`). Archived rows are handled by the tombstone pass below.
-    let live: Vec<&_> = rows
-        .iter()
-        .filter(|r| !r.archived && r.remote_id.is_none())
-        .collect();
-    let already_synced = rows
-        .iter()
-        .filter(|r| !r.archived && r.remote_id.is_some())
-        .count();
-    let attempted = live.len();
-    // Progress is only worth emitting when the push actually spans multiple
-    // chunks; a single-chunk push stays quiet (no noise on small pushes).
-    let multi_chunk = attempted.div_ceil(PUSH_BATCH_CHUNK_SIZE) > 1;
-    // Map external_id (local uuid) → local_id so we can record the cloud-minted
-    // id returned in the 207 result back onto the local row.
-    for chunk in live.chunks(PUSH_BATCH_CHUNK_SIZE) {
-        let mut items: Vec<BatchPushItem> = Vec::with_capacity(chunk.len());
-        for r in chunk {
-            // Only read the local embedding when the server can accept it. The
-            // stored blob is raw little-endian fp32 (`vec_to_blob`); decode it
-            // and only attach a correctly-dimensioned (896) vector — a
-            // wrong-length or missing embedding falls back to text-only rather
-            // than poisoning the whole batch with a 4xx.
-            let vector = if accepts_pushed_vectors {
-                local
-                    .get_embedding(r.local_id)?
-                    .map(|blob| spelunk_core::embeddings::blob_to_vec(&blob))
-                    .filter(|v| v.len() == spelunk_core::embeddings::EMBEDDING_DIM)
-            } else {
-                None
-            };
-            items.push(
-                BatchPushItem {
-                    kind: r.kind.clone(),
-                    title: r.title.clone(),
-                    body: if r.body.is_empty() {
-                        None
-                    } else {
-                        Some(r.body.clone())
-                    },
-                    external_id: r.uuid.clone(),
-                    source_commit: r.source_ref.clone(),
-                    vector: None,
-                    vector_model: None,
-                    vector_precision: None,
-                }
-                .maybe_attach_vector(accepts_pushed_vectors, vector),
-            );
-        }
-
-        // A chunk failure (timeout / transport / non-2xx) usually means the
-        // server is overloaded; pushing the remaining chunks would only make that
-        // worse, and the resumable design (already-stamped chunks are filtered
-        // out of `live` on the next run) means a re-run resumes from exactly
-        // here. So stop at the first failed chunk and return the progress that
-        // already landed instead of discarding it via `?`.
-        let res = match client.push_batch(items).await {
-            Ok(res) => res,
-            Err(e) => {
-                interrupted = Some(e.to_string());
-                break;
-            }
-        };
-
-        // `created`/`skipped`/`failed` (aggregate ints) and `results[]`
-        // (per-item) are independent fields on `BatchPushResult` — nothing on
-        // the wire guarantees they agree, and a server can send an aggregate
-        // `created: 0` for a batch whose `results[]` shows the entries
-        // durably persisted. The aggregate ints are NOT trusted here: tally
-        // from `results[].status`, the reconciled signal, and only fall back
-        // to the aggregate when the server sent no per-item detail at all to
-        // reconcile against.
-        if res.results.is_empty() {
-            created += res.created;
-            skipped += res.skipped;
-            failed += res.failed;
-        }
-
-        // Record cloud ids for created entries so a later pull dedupes them and
-        // a later archive can tombstone them by id.
-        for item in &res.results {
-            match item.status.as_str() {
-                "created" => created += 1,
-                "skipped" => skipped += 1,
-                // Anything else — `"failed"`, or an unrecognized status — did
-                // not affirmatively land; count it as failed rather than
-                // silently dropping it from every tally.
-                _ => failed += 1,
-            }
-            // Stamping `remote_id` is permanent (it's what excludes a row from
-            // `live` on every future push), so only do it for a status that
-            // affirmatively means the cloud durably has this row: `created`
-            // (just persisted) or `skipped` (already persisted — dedup on
-            // identity). Any other status — `failed`, or an id riding along
-            // with a status that doesn't mean persisted — must not stamp, or
-            // that row can never be retried again.
-            let durably_persisted = item.status == "created" || item.status == "skipped";
-            if durably_persisted
-                && let (Some(ext), Some(cloud_id)) =
-                    (item.external_id.as_deref(), item.id.as_deref())
-                && let Some(row) = chunk.iter().find(|r| r.uuid == ext)
-            {
-                local.set_remote_id(row.local_id, cloud_id)?;
-            }
-            if item.status == "failed" {
-                eprintln!(
-                    "  [push-fail] {}",
-                    item.external_id.as_deref().unwrap_or("<unknown>")
-                );
-            }
-        }
-
-        if multi_chunk {
-            // `done` is the running durably-landed count (created + skipped),
-            // which never exceeds `attempted`; the final `done` equals
-            // `created + skipped`.
-            on_progress((created + skipped) as usize, attempted);
-        }
-    }
-
-    // Tombstone archived entries that the cloud already knows about. An archived
-    // row with no `remote_id` was never pushed live, so there is nothing to
-    // delete cloud-side; we skip it. Skipped entirely on an interrupted push: the
-    // connection is already failing and the remaining live chunks were not sent.
-    if interrupted.is_none() && include_archived {
-        for r in rows.iter().filter(|r| r.archived) {
-            if let Some(remote_id) = r.remote_id.as_deref() {
-                client.delete_remote(remote_id).await?;
-            }
-        }
-    }
-
-    Ok(PushSummary {
-        attempted,
-        created,
-        skipped,
-        failed,
-        already_synced,
-        interrupted,
-    })
-}
-
-/// Pull remote entries after the UUID cursor and apply them idempotently.
-/// Returns the number of newly-inserted local rows.
-///
-/// The cursor is derived from the store itself — `MAX(remote_id)` over local
-/// notes (decision #183) — so there is no persisted watermark to advance: the
-/// next run re-derives the cursor from the rows just applied. This is what makes
-/// the pull immune to clock drift and trivially resumable. Used as-is by the
-/// one-way `memory pull`; `sync_round` below instead calls
-/// [`pull_and_apply_since`] directly so it can pin an explicit cursor across
-/// its two pull passes.
-async fn pull_and_apply(local: &MemoryStore, client: &CloudSyncClient) -> Result<usize> {
-    let cursor = local.max_remote_id()?;
-    pull_and_apply_since(local, client, cursor.as_deref()).await
-}
-
-/// Pull remote entries after an explicit `since_id` cursor and apply them
-/// idempotently. Returns the number of newly-inserted local rows.
-///
-/// `pull_since` returns at most `CloudSyncClient::PULL_PAGE_LIMIT` entries
-/// per call, so a backlog larger than one page requires more than one
-/// request: this loops, applying each page as it arrives and advancing the
-/// cursor to the last entry's `id`, until a page comes back shorter than the
-/// page limit (the definitive "nothing left" signal — including the
-/// empty-page case for an already-fully-synced project). Without this loop,
-/// a first sync into an established project would silently apply only the
-/// first page and report success.
-///
-/// Applying an entry is idempotent regardless of how many times this is
-/// called with overlapping ranges, or how many pages one call fetches:
-/// [`MemoryStore::apply_remote_note`] dedupes on `remote_id` (or reuses a
-/// matching row by `entity_id`), so re-fetching a row already known locally
-/// is a harmless no-op, not a duplicate insert or a double count.
-async fn pull_and_apply_since(
-    local: &MemoryStore,
-    client: &CloudSyncClient,
-    cursor: Option<&str>,
-) -> Result<usize> {
-    let mut cursor = cursor.map(str::to_string);
-    let mut applied = 0usize;
-    loop {
-        let entries = client.pull_since(cursor.as_deref()).await?;
-        let page_len = entries.len();
-
-        for e in &entries {
-            let created_secs = parse_iso_to_secs(&e.created_at);
-            let inserted = local.apply_remote_note(
-                &e.id,
-                &e.kind,
-                &e.title,
-                e.body.as_deref().unwrap_or(""),
-                e.source_commit.as_deref(),
-                created_secs,
-                e.is_archived(),
-            )?;
-            if inserted {
-                applied += 1;
-            }
-        }
-
-        if (page_len as i64) < CloudSyncClient::PULL_PAGE_LIMIT {
-            break;
-        }
-        // A full page never proves it's the last one (the server never
-        // returns more than the limit even when more remain), so always
-        // follow up: entries is non-empty here (page_len == the limit, which
-        // is > 0), so `last()` is always `Some`.
-        cursor = entries.last().map(|e| e.id.clone());
-    }
-    Ok(applied)
-}
-
-/// Outcome of one [`sync_round`]: the push summary plus the total newly
-/// applied entries across both pull passes.
-#[derive(Debug)]
-struct SyncRoundOutcome {
-    pushed: PushSummary,
-    pulled: usize,
-}
-
-/// Run one full two-way sync round: pull, then push, then pull again off the
-/// same pre-round cursor.
-///
-/// This is `spelunk sync`'s actual push+pull sequence, extracted into its own
-/// function so it can be exercised directly in tests against a real server:
-/// the command entry point (`memory_sync`) can't be unit-tested cheaply
-/// because of its config/tier-probe plumbing (`get_tier`'s per-process cache
-/// makes multiple differently-configured in-process probes unreliable within
-/// one test binary).
-///
-/// Neither a plain push-then-pull nor a plain pull-then-push reorder is
-/// sufficient here. The failure mode: the cursor is always `MAX(remote_id)`
-/// over local rows (decision #183, no persisted watermark), and this
-/// client's own push mints `remote_id`s stamped "now", chronologically the
-/// newest thing on the server. If a plain re-derived cursor were used for a
-/// second pull, this round's own just-pushed rows would become the new
-/// `MAX(remote_id)`, permanently shadowing (via the strict `>` comparison)
-/// any teammate entry that landed between this round's first pull and its
-/// own push.
-///
-/// The fix: capture the cursor once, before this round's own pull or push
-/// touches anything (`pre_round_cursor`), pull with it, push, then pull AGAIN
-/// reusing that SAME `pre_round_cursor`, not a freshly re-derived one. The
-/// second pull harmlessly re-fetches this round's own just-pushed rows (their
-/// `remote_id` is now `> pre_round_cursor`) alongside anything a teammate
-/// pushed in the gap; both are idempotent no-ops or genuine new applies via
-/// [`pull_and_apply_since`], so the combined count is never inflated by
-/// double-counting.
-async fn sync_round(
-    local: &MemoryStore,
-    client: &CloudSyncClient,
-    include_archived: bool,
-    accepts_pushed_vectors: bool,
-) -> Result<SyncRoundOutcome> {
-    let pre_round_cursor = local.max_remote_id()?;
-
-    let pulled_first = pull_and_apply_since(local, client, pre_round_cursor.as_deref()).await?;
-
-    let pushed = push_local(local, client, include_archived, accepts_pushed_vectors).await?;
-
-    // If this second pull errors (network blip, transient 5xx), the error
-    // propagates out of `sync_round` rather than being swallowed — `?`
-    // surfaces it to `memory_sync`, which reports a failure and a non-zero
-    // exit. That is correct (a real error must not be silently dropped), but
-    // by this point `pushed` already reflects a push that may have durably
-    // landed server-side (and stamped local `remote_id`s accordingly, inside
-    // `push_local`, before this call even runs) — the failure is scoped to
-    // the confirmation pull, not the push. Attach that context so the
-    // surfaced error doesn't read as "nothing happened": a caller shouldn't
-    // conclude their content was lost and try to force a re-push (harmless
-    // but pointless — already-stamped rows are excluded from `live` and
-    // skipped) instead of simply re-running sync, which retries the pull with
-    // an unaffected, freshly-derived cursor.
-    let pulled_second = pull_and_apply_since(local, client, pre_round_cursor.as_deref())
-        .await
-        .with_context(|| {
-            format!(
-                "confirmation pull failed after this round's push already reached \
-                 the server ({} attempted: {} created, {} skipped, {} failed) — \
-                 the push is not affected by this error; re-running sync will retry \
-                 the pull without re-pushing already-landed entries",
-                pushed.attempted, pushed.created, pushed.skipped, pushed.failed
-            )
-        })?;
-
-    Ok(SyncRoundOutcome {
-        pushed,
-        pulled: pulled_first + pulled_second,
-    })
-}
-
-/// Parse an ISO 8601 / RFC 3339 timestamp to Unix epoch seconds.
-///
-/// Falls back to "now" if the server sends a value we cannot parse, so a single
-/// odd row never aborts the whole sync.
-pub(super) fn parse_iso_to_secs(s: &str) -> i64 {
-    chrono::DateTime::parse_from_rfc3339(s)
-        .map(|dt| dt.timestamp())
-        .unwrap_or_else(|_| crate::storage::now_secs())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_iso_to_secs_handles_utc_z() {
-        // 2021-01-01T00:00:00Z = 1609459200
-        assert_eq!(parse_iso_to_secs("2021-01-01T00:00:00Z"), 1_609_459_200);
-    }
-
-    #[test]
-    fn parse_iso_to_secs_handles_offset() {
-        // 2021-01-01T01:00:00+01:00 == 2021-01-01T00:00:00Z
-        assert_eq!(
-            parse_iso_to_secs("2021-01-01T01:00:00+01:00"),
-            1_609_459_200
-        );
-    }
-
-    #[test]
-    fn parse_iso_to_secs_falls_back_on_garbage() {
-        // Must not panic; returns some positive epoch (now).
-        assert!(parse_iso_to_secs("not-a-timestamp") > 0);
-    }
-
-    // ── resolve_sync_project ───────────────────────────────────────────────
-    // Sync must never invent a project name. With neither `--project` nor a
-    // configured `project_id`, the call halts with a message pointing the user
-    // at `--project <slug>`; with an explicit slug (or configured id), that slug
-    // is threaded through verbatim so it reaches the outbound request.
-
-    fn cfg_with_project(id: Option<&str>) -> Config {
-        Config {
-            project_id: id.map(str::to_string),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn resolve_sync_project_halts_when_nothing_configured_or_passed() {
-        let cfg = cfg_with_project(None);
-        let err = resolve_sync_project(None, &cfg).unwrap_err();
-        let msg = err.to_string();
-        // Actionable: names the exact re-run and refuses to guess.
-        assert!(msg.contains("--project <slug>"), "msg: {msg}");
-        assert!(
-            msg.contains("never guessed") || msg.contains("git remote"),
-            "must state it won't auto-derive: {msg}"
-        );
-    }
-
-    #[test]
-    fn resolve_sync_project_uses_cli_flag_when_passed() {
-        let cfg = cfg_with_project(None);
-        let slug = resolve_sync_project(Some("acme/app"), &cfg).unwrap();
-        assert_eq!(slug, "acme/app");
-    }
-
-    #[test]
-    fn resolve_sync_project_falls_back_to_configured_id() {
-        let cfg = cfg_with_project(Some("team/proj"));
-        let slug = resolve_sync_project(None, &cfg).unwrap();
-        assert_eq!(slug, "team/proj");
-    }
-
-    #[test]
-    fn resolve_sync_project_cli_flag_overrides_configured_id() {
-        let cfg = cfg_with_project(Some("team/proj"));
-        let slug = resolve_sync_project(Some("other/slug"), &cfg).unwrap();
-        assert_eq!(slug, "other/slug");
-    }
-
-    #[test]
-    fn resolve_sync_project_treats_blank_slug_as_absent() {
-        // A whitespace-only `--project ""` must not silently pass an empty slug.
-        let cfg = cfg_with_project(None);
-        assert!(resolve_sync_project(Some("   "), &cfg).is_err());
-    }
-
-    // ── end-to-end first-run path ──────────────────────────────────────────
-    // The story's target path: a first-run user has a non-loopback team
-    // `server_url`, NO configured `project_id`, and passes `--project <slug>`.
-    // Before the fix this was rejected at dispatch by `cfg.validate()` before
-    // `resolve_sync_project` ever ran. This test walks the same two gates the
-    // dispatcher + `memory_sync` cross — (1) config validation must accept the
-    // `--project`-only config, (2) the resolved slug must reach the outbound
-    // request — proving the path is live end to end (minus the auth/tier
-    // machinery, which is orthogonal to this story).
-    #[tokio::test]
-    async fn first_run_project_flag_only_passes_dispatch_and_reaches_wire() {
-        use crate::storage::{BatchPushItem, CloudSyncClient};
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        // A non-loopback team server_url with NO project_id — a genuine first run.
-        let cfg = Config {
-            server_url: Some("http://spelunk.internal:7777".to_string()),
-            project_id: None,
-            ..Default::default()
-        };
-        let cli_project = Some("acme/app");
-
-        // Gate 1 — dispatch: `--project` makes a project available, so the
-        // non-loopback server_url no longer blocks (regression under test).
-        let project_available = cli_project.is_some() || cfg.project_id.is_some();
-        cfg.validate_with_project(project_available)
-            .expect("first-run --project must pass dispatch validation");
-
-        // Gate 2 — resolution: the explicit slug wins and is what sync targets.
-        let slug = resolve_sync_project(cli_project, &cfg).unwrap();
-        assert_eq!(slug, "acme/app");
-
-        // Wire: the resolved slug must land, percent-encoded, in the request
-        // path so the server can lazily create/reuse that project.
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/acme%2Fapp/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 0,
-                "results": [{"status": "created", "external_id": "e1", "id": "cloud-1"}]
-            })))
-            .mount(&server)
-            .await;
-
-        let client = CloudSyncClient::new(&server.uri(), &slug, None, None).unwrap();
-        let res = client
-            .push_batch(vec![BatchPushItem {
-                kind: "decision".into(),
-                title: "T".into(),
-                body: Some("B".into()),
-                external_id: "e1".into(),
-                source_commit: None,
-                vector: None,
-                vector_model: None,
-                vector_precision: None,
-            }])
-            .await
-            .expect("push to the lazily-created project must succeed");
-        assert_eq!(res.created, 1);
-    }
-
-    // ── push_local end-to-end: remote_id stamping + idempotent re-sync ─────
-    // The local-first push path is where the server-minted
-    // cross-machine id is PERSISTED — stamped onto `notes.remote_id` from the
-    // 207 batch result — not the `RemoteMemoryBackend::add` debug-log path
-    // (which is the cloud-first, remote-is-store-of-record case with no local
-    // row). Locks in that a push stamps `remote_id` and a re-push sends nothing
-    // (no duplicate cloud writes, no local dupes).
-
-    fn register_sqlite_vec() {
-        use std::sync::OnceLock;
-        // `MemoryStore::open` creates a vec0 table, so the extension must be
-        // registered before any connection opens.
-        static INIT: OnceLock<()> = OnceLock::new();
-        INIT.get_or_init(|| {
-            #[allow(clippy::missing_transmute_annotations)]
-            unsafe {
-                rusqlite::ffi::sqlite3_auto_extension(Some(std::mem::transmute(
-                    sqlite_vec::sqlite3_vec_init as *const (),
-                )));
-            }
-        });
-    }
-
-    // ── chunking, resumability, and progress (D1 / D4 / D5) ─────────────────
-
-    use std::collections::{HashMap, HashSet};
-    use tempfile::TempDir;
-    use wiremock::matchers::{method, path};
-    use wiremock::{Mock, MockServer, Request, Respond, ResponseTemplate};
-
-    // Seed `n` distinct live notes: distinct titles give distinct entity_ids (no
-    // dedupe collision), and each is unstamped so all `n` land in the push set.
-    fn seed_notes(store: &MemoryStore, n: usize) {
-        for i in 0..n {
-            store
-                .add_note("note", &format!("T{i}"), "body", &[], &[], None, None)
-                .unwrap();
-        }
-    }
-
-    // Echoes every received entry back as `created` with a distinct cloud id, so
-    // `push_local` tallies and stamps exactly as a real 207 would, for any chunk
-    // size (a static body cannot, since the ids it must echo are minted per row).
-    struct EchoCreated;
-    impl Respond for EchoCreated {
-        fn respond(&self, request: &Request) -> ResponseTemplate {
-            let body: serde_json::Value = serde_json::from_slice(&request.body).unwrap_or_default();
-            let results: Vec<serde_json::Value> = body["entries"]
-                .as_array()
-                .map(|entries| {
-                    entries
-                        .iter()
-                        .map(|e| {
-                            let ext = e["external_id"].as_str().unwrap_or_default();
-                            serde_json::json!({
-                                "status": "created",
-                                "external_id": ext,
-                                "id": format!("cloud-{ext}"),
-                            })
-                        })
-                        .collect()
-                })
-                .unwrap_or_default();
-            ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": results.len(), "skipped": 0, "failed": 0, "results": results
-            }))
-        }
-    }
-
-    // First request lands (echoed created), every later request 500s. Models a
-    // server that accepts the first chunk then falls over. Keyed on call count so
-    // it does not depend on wiremock's ordering of same-path mocks.
-    struct FailAfterFirst {
-        calls: std::sync::Arc<std::sync::atomic::AtomicUsize>,
-    }
-    impl Respond for FailAfterFirst {
-        fn respond(&self, request: &Request) -> ResponseTemplate {
-            let n = self.calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-            if n == 0 {
-                EchoCreated.respond(request)
-            } else {
-                ResponseTemplate::new(500).set_body_string("overloaded")
-            }
-        }
-    }
-
-    // Echoes `created` for the first `ok` requests, then 500s every later one.
-    // Unlike `FailAfterFirst`, the failure can be placed on any chunk, so a test
-    // can prove the interrupted summary counts every chunk that landed before the
-    // failure and that the loop halts exactly at the failed chunk.
-    struct FailAfterN {
-        ok: usize,
-        calls: std::sync::Arc<std::sync::atomic::AtomicUsize>,
-    }
-    impl Respond for FailAfterN {
-        fn respond(&self, request: &Request) -> ResponseTemplate {
-            let n = self.calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-            if n < self.ok {
-                EchoCreated.respond(request)
-            } else {
-                ResponseTemplate::new(500).set_body_string("overloaded")
-            }
-        }
-    }
-
-    #[test]
-    fn push_batch_chunk_size_constant_is_50() {
-        assert_eq!(PUSH_BATCH_CHUNK_SIZE, 50);
-    }
-
-    #[test]
-    fn chunking_splits_live_set_on_the_constant() {
-        let n = PUSH_BATCH_CHUNK_SIZE;
-        // (total, expected chunk count, expected last-chunk len).
-        let cases = [
-            (0usize, 0usize, 0usize),
-            (1, 1, 1),
-            (n, 1, n),
-            (n + 1, 2, 1),
-            (3 * n, 3, n),
-            (2 * n + 7, 3, 7),
-        ];
-        for (total, want_chunks, want_last) in cases {
-            let items: Vec<usize> = (0..total).collect();
-            let chunks: Vec<&[usize]> = items.chunks(PUSH_BATCH_CHUNK_SIZE).collect();
-            assert_eq!(chunks.len(), want_chunks, "chunk count for total={total}");
-            assert!(
-                chunks.iter().all(|c| c.len() <= PUSH_BATCH_CHUNK_SIZE),
-                "no chunk exceeds N for total={total}"
-            );
-            if let Some((last, rest)) = chunks.split_last() {
-                assert!(
-                    rest.iter().all(|c| c.len() == PUSH_BATCH_CHUNK_SIZE),
-                    "every chunk but the last is exactly N for total={total}"
-                );
-                assert_eq!(last.len(), want_last, "last-chunk len for total={total}");
-            }
-            let flat: Vec<usize> = chunks.iter().flat_map(|c| c.iter().copied()).collect();
-            assert_eq!(
-                flat, items,
-                "chunks cover every entry exactly once for total={total}"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn push_chunks_cover_every_entry_exactly_once() {
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 120); // ceil(120 / 50) = 3 chunks
-
-        let want: HashSet<String> = store
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.uuid)
-            .collect();
-        assert_eq!(want.len(), 120);
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(EchoCreated)
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!(
-            (s.attempted, s.created, s.skipped, s.failed),
-            (120, 120, 0, 0)
-        );
-        assert!(s.interrupted.is_none());
-
-        let reqs = server.received_requests().await.unwrap();
-        assert_eq!(reqs.len(), 3, "ceil(120 / 50) POSTs");
-        let mut seen: Vec<String> = Vec::new();
-        let mut sizes: Vec<usize> = Vec::new();
-        for req in &reqs {
-            let body: serde_json::Value = serde_json::from_slice(&req.body).unwrap();
-            let entries = body["entries"].as_array().unwrap();
-            sizes.push(entries.len());
-            for e in entries {
-                seen.push(e["external_id"].as_str().unwrap().to_string());
-            }
-        }
-        // A `<= N` check alone would pass an implementation chunking on the wrong
-        // size (e.g. 40), so pin the exact per-request counts: full chunks up to
-        // the constant, then the remainder. Requests arrive in push order.
-        assert_eq!(sizes, vec![50, 50, 20], "exact per-request entry counts");
-        assert_eq!(seen.len(), 120, "no entry is pushed more than once");
-        assert_eq!(
-            seen.into_iter().collect::<HashSet<_>>(),
-            want,
-            "the union of pushed ids equals the full live set"
-        );
-    }
-
-    #[tokio::test]
-    async fn multi_chunk_push_threads_slug_into_every_request_path() {
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 60); // 2 chunks
-
-        let want: HashSet<String> = store
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.uuid)
-            .collect();
-
-        // The mock ONLY matches the percent-encoded slug path, so an all-created
-        // push proves every chunk (the first included) carried the slug in the
-        // path, which is what lazily creates/reuses the project server-side.
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/acme%2Fapp/memory/batch"))
-            .respond_with(EchoCreated)
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "acme/app", None, None).unwrap();
-
-        let s = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s.attempted, s.created), (60, 60));
-        assert!(s.interrupted.is_none());
-
-        let reqs = server.received_requests().await.unwrap();
-        assert_eq!(reqs.len(), 2);
-        let seen: HashSet<String> = reqs
-            .iter()
-            .flat_map(|r| {
-                let body: serde_json::Value = serde_json::from_slice(&r.body).unwrap();
-                body["entries"]
-                    .as_array()
-                    .unwrap()
-                    .iter()
-                    .map(|e| e["external_id"].as_str().unwrap().to_string())
-                    .collect::<Vec<_>>()
-            })
-            .collect();
-        assert_eq!(seen, want, "all entries observed across the batch requests");
-    }
-
-    #[tokio::test]
-    async fn interrupted_push_stops_and_resumes_from_the_remainder() {
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 120); // 3 chunks: 50, 50, 20
-
-        // ── Run 1: chunk 1 lands, chunk 2 fails (500), chunk 3 is never sent. ──
-        let server1 = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(FailAfterFirst {
-                calls: std::sync::Arc::new(std::sync::atomic::AtomicUsize::new(0)),
-            })
-            .mount(&server1)
-            .await;
-        let client1 = CloudSyncClient::new(&server1.uri(), "proj", None, None).unwrap();
-
-        let s1 = push_local(&store, &client1, false, false).await.unwrap();
-        assert!(
-            s1.interrupted.is_some(),
-            "a mid-push failure marks the summary interrupted"
-        );
-        assert_eq!(
-            (s1.attempted, s1.created, s1.skipped),
-            (120, 50, 0),
-            "only the first chunk landed"
-        );
-        assert_eq!(
-            server1.received_requests().await.unwrap().len(),
-            2,
-            "must stop at the first failed chunk, never push the remaining chunk"
-        );
-        let remaining: Vec<_> = store
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .filter(|r| r.remote_id.is_none())
-            .collect();
-        assert_eq!(
-            remaining.len(),
-            70,
-            "the 50 landed rows are durably stamped"
-        );
-
-        // ── Run 2: a healthy server pushes ONLY the remainder. ──
-        let server2 = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(EchoCreated)
-            .mount(&server2)
-            .await;
-        let client2 = CloudSyncClient::new(&server2.uri(), "proj", None, None).unwrap();
-
-        let s2 = push_local(&store, &client2, false, false).await.unwrap();
-        assert!(s2.interrupted.is_none());
-        assert_eq!(
-            (s2.attempted, s2.created),
-            (70, 70),
-            "resume pushes only the remainder"
-        );
-        let reqs2 = server2.received_requests().await.unwrap();
-        let pushed2: usize = reqs2
-            .iter()
-            .map(|r| {
-                let body: serde_json::Value = serde_json::from_slice(&r.body).unwrap();
-                body["entries"].as_array().unwrap().len()
-            })
-            .sum();
-        assert_eq!(
-            pushed2, 70,
-            "resume re-pushes exactly the previously-unstamped rows"
-        );
-        assert!(
-            store
-                .rows_for_sync(false)
-                .unwrap()
-                .iter()
-                .all(|r| r.remote_id.is_some()),
-            "every row is stamped after the resume"
-        );
-        assert_eq!(
-            store.count().unwrap(),
-            120,
-            "no local duplicates introduced"
-        );
-    }
-
-    #[tokio::test]
-    async fn interrupted_on_a_later_chunk_counts_only_the_chunks_that_landed() {
-        // The failure lands on chunk 3 of 7: `created` must equal exactly the two
-        // full chunks that landed before it (not one, not three), and the loop
-        // must issue no request past the failed chunk.
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 340); // ceil(340 / 50) = 7 chunks
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(FailAfterN {
-                ok: 2,
-                calls: std::sync::Arc::new(std::sync::atomic::AtomicUsize::new(0)),
-            })
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s = push_local(&store, &client, false, false).await.unwrap();
-        assert!(s.interrupted.is_some());
-        assert_eq!(
-            (s.attempted, s.created, s.skipped),
-            (340, 100, 0),
-            "created reflects exactly the two chunks that landed before the failure"
-        );
-        assert_eq!(
-            server.received_requests().await.unwrap().len(),
-            3,
-            "halts on the failed 3rd chunk; chunks 4 through 7 are never sent"
-        );
-        let unstamped = store
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .filter(|r| r.remote_id.is_none())
-            .count();
-        assert_eq!(unstamped, 240, "only the 100 landed rows are stamped");
-    }
-
-    #[tokio::test]
-    async fn interrupted_push_skips_the_tombstone_delete_pass() {
-        // Once a live chunk fails the connection is already failing, so the
-        // archived-entry DELETEs must not be issued either. `delete_remote` treats
-        // a 404 as success, so only a request-count check catches a regression
-        // that dropped the interrupted gate on the tombstone pass.
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        // Three live, unstamped rows form the single push chunk that will fail.
-        seed_notes(&store, 3);
-        // Two archived rows that DO carry a remote_id: the tombstone pass would
-        // DELETE these if it were not skipped on an interrupted push.
-        for tag in ["A0", "A1"] {
-            let (id, _) = store
-                .add_note("note", tag, "body", &[], &[], None, None)
-                .unwrap();
-            store.set_remote_id(id, &format!("cloud-{tag}")).unwrap();
-            assert!(store.archive(id).unwrap());
-        }
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(500).set_body_string("overloaded"))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s = push_local(&store, &client, true, false).await.unwrap();
-        assert!(s.interrupted.is_some());
-        assert_eq!(s.created, 0, "the only live chunk failed");
-
-        let reqs = server.received_requests().await.unwrap();
-        assert_eq!(
-            reqs.len(),
-            1,
-            "only the failing batch POST is sent; no tombstone DELETEs follow"
-        );
-        assert!(
-            reqs[0].url.path().ends_with("/memory/batch"),
-            "the single request is the batch push, not a delete"
-        );
-    }
-
-    #[tokio::test]
-    async fn overlapping_repush_tallies_skipped_and_leaves_no_duplicates() {
-        // Models the committed-but-unstamped overlap: a prior push persisted rows
-        // server-side but the client lost the response (no `remote_id` stamped).
-        // On re-push the server dedupes on external_id and returns 207 `skipped`
-        // for the overlap; the client must tally those as skipped (not created),
-        // stamp them, and add no local duplicates.
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 4);
-
-        let rows = store.rows_for_sync(false).unwrap();
-        // First two rows are the overlap (already server-side → skipped); the last
-        // two are genuinely new (created). Keyed by uuid, so row order is irrelevant.
-        let mut status: HashMap<String, &str> = HashMap::new();
-        status.insert(rows[0].uuid.clone(), "skipped");
-        status.insert(rows[1].uuid.clone(), "skipped");
-        status.insert(rows[2].uuid.clone(), "created");
-        status.insert(rows[3].uuid.clone(), "created");
-        let results: Vec<serde_json::Value> = rows
-            .iter()
-            .map(|r| {
-                serde_json::json!({
-                    "status": status[&r.uuid],
-                    "external_id": r.uuid,
-                    "id": format!("cloud-{}", r.uuid),
-                })
-            })
-            .collect();
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 2, "skipped": 2, "failed": 0, "results": results
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s.attempted, s.created, s.skipped, s.failed), (4, 2, 2, 0));
-        assert!(s.interrupted.is_none());
-        assert!(
-            store
-                .rows_for_sync(false)
-                .unwrap()
-                .iter()
-                .all(|r| r.remote_id.is_some()),
-            "overlap and new rows both end stamped"
-        );
-        assert_eq!(
-            store.count().unwrap(),
-            4,
-            "no local duplicates from the round trip"
-        );
-
-        let s2 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s2.attempted, s2.already_synced), (0, 4));
-        assert_eq!(
-            server.received_requests().await.unwrap().len(),
-            1,
-            "a fully-stamped re-push sends no batch request"
-        );
-    }
-
-    #[tokio::test]
-    async fn multi_chunk_push_reports_cumulative_progress_after_each_chunk() {
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, 120); // 3 chunks
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(EchoCreated)
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let mut progress: Vec<(usize, usize)> = Vec::new();
-        let s = push_local_reporting(&store, &client, false, false, |done, total| {
-            progress.push((done, total));
-        })
-        .await
-        .unwrap();
-
-        assert_eq!(progress.len(), 3, "one progress emission per chunk");
-        assert!(
-            progress.iter().all(|&(done, total)| done <= total),
-            "done never exceeds total: {progress:?}"
-        );
-        assert!(
-            progress.windows(2).all(|w| w[0].0 <= w[1].0),
-            "cumulative done never regresses across chunks: {progress:?}"
-        );
-        assert_eq!(
-            progress.iter().map(|&(d, _)| d).collect::<Vec<_>>(),
-            vec![50, 100, 120],
-            "done advances cumulatively by the landed count"
-        );
-        assert!(progress.iter().all(|&(_, total)| total == 120));
-        assert_eq!(
-            progress.last().unwrap().0 as u32,
-            s.created + s.skipped,
-            "final reported done equals created + skipped"
-        );
-        assert_eq!(s.created + s.skipped, 120);
-    }
-
-    #[tokio::test]
-    async fn single_chunk_push_is_one_request_and_emits_no_progress() {
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        seed_notes(&store, PUSH_BATCH_CHUNK_SIZE); // exactly one chunk
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(EchoCreated)
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let mut calls = 0usize;
-        let s = push_local_reporting(&store, &client, false, false, |_, _| calls += 1)
-            .await
-            .unwrap();
-        assert_eq!(calls, 0, "a single-chunk push emits no progress");
-        assert_eq!(
-            server.received_requests().await.unwrap().len(),
-            1,
-            "exactly one POST for a push of N or fewer entries"
-        );
-        assert_eq!(s.created, PUSH_BATCH_CHUNK_SIZE as u32);
-        assert!(
-            s.interrupted.is_none(),
-            "a clean single-chunk push is never marked interrupted"
-        );
-    }
-
-    #[tokio::test]
-    async fn push_local_stamps_remote_id_and_repush_is_idempotent() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        store
-            .add_note("note", "Two", "second", &[], &[], None, None)
-            .unwrap();
-
-        // Learn the lazily-minted external_ids up front so the mock can echo
-        // them back with distinct cloud ids; `ensure_uuid` is idempotent, so the
-        // push below re-derives the same uuids.
-        let rows = store.rows_for_sync(false).unwrap();
-        assert_eq!(rows.len(), 2);
-        let (ext_a, ext_b) = (rows[0].uuid.clone(), rows[1].uuid.clone());
-        let cloud_a = "01890000-0000-7000-8000-0000000000a1";
-        let cloud_b = "01890000-0000-7000-8000-0000000000a2";
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 2, "skipped": 0, "failed": 0,
-                "results": [
-                    {"status": "created", "external_id": ext_a, "id": cloud_a},
-                    {"status": "created", "external_id": ext_b, "id": cloud_b},
-                ]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        // First push: creates both, persists the server-minted id on each row.
-        let s1 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s1.attempted, s1.created, s1.skipped), (2, 2, 0));
-        assert_eq!(
-            store.note_id_for_remote_id(cloud_a).unwrap(),
-            Some(rows[0].local_id)
-        );
-        assert_eq!(
-            store.note_id_for_remote_id(cloud_b).unwrap(),
-            Some(rows[1].local_id)
-        );
-        // The pull cursor is now the newest stamped id.
-        assert_eq!(store.max_remote_id().unwrap().as_deref(), Some(cloud_b));
-
-        // Second push: every row carries a `remote_id`, so the live set is empty
-        // and no batch request is sent — the re-sync is a no-op. `attempted` must
-        // reflect that (not the raw row count), so callers never report "Pushed
-        // N" when nothing was sent.
-        let s2 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s2.attempted, s2.created, s2.already_synced), (0, 0, 2));
-        assert_eq!(
-            server.received_requests().await.unwrap().len(),
-            1,
-            "re-push must not hit the batch endpoint again"
-        );
-        // No duplicate local rows introduced by the round trip.
-        assert_eq!(store.count().unwrap(), 2);
-    }
-
-    // ── pushed-vector fast path ─────────────────────────────────────────────
-    // A note with a local fp32/896 embedding carries that vector (+ model tag
-    // + precision "fp32") to a server advertising `accepts_pushed_vectors`, so
-    // the server stores it as-is; against a server without the capability the
-    // same note is pushed text-only even though the vector is available. This
-    // exercises the full `push_local` wiring: it reads the local embedding and
-    // consults the gate, which the `maybe_attach_vector` unit test cannot.
-
-    /// Insert an active note plus a valid L2-normalised fp32/896 embedding,
-    /// returning its local id + external uuid.
-    fn note_with_embedding(store: &MemoryStore) -> (i64, String) {
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        let dim = spelunk_core::embeddings::EMBEDDING_DIM;
-        let vec: Vec<f32> = vec![1.0 / (dim as f32).sqrt(); dim];
-        let blob = spelunk_core::embeddings::vec_to_blob(&vec);
-        let rows = store.rows_for_sync(false).unwrap();
-        assert_eq!(rows.len(), 1);
-        store.insert_embedding(rows[0].local_id, &blob).unwrap();
-        (rows[0].local_id, rows[0].uuid.clone())
-    }
-
-    #[tokio::test]
-    async fn push_local_attaches_vector_when_server_accepts() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        let (_id, uuid) = note_with_embedding(&store);
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 0,
-                "results": [{"status": "created", "external_id": uuid, "id": "cloud-1"}]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        // accepts_pushed_vectors = true → the fp32/896 vector reaches the wire.
-        push_local(&store, &client, false, true).await.unwrap();
-
-        let reqs = server.received_requests().await.unwrap();
-        let json: serde_json::Value = serde_json::from_slice(&reqs[0].body).unwrap();
-        let entry = &json["entries"][0];
-        assert_eq!(
-            entry["vector"].as_array().map(Vec::len),
-            Some(spelunk_core::embeddings::EMBEDDING_DIM),
-            "server that accepts vectors must receive the 896-dim vector: {entry}"
-        );
-        assert_eq!(entry["vector_model"], "F2LLM-v2-330M");
-        assert_eq!(entry["vector_precision"], "fp32");
-    }
-
-    #[tokio::test]
-    async fn push_local_stays_text_only_when_server_declines() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        let (_id, uuid) = note_with_embedding(&store);
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 0,
-                "results": [{"status": "created", "external_id": uuid, "id": "cloud-1"}]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        // accepts_pushed_vectors = false → text-only, despite a local vector.
-        push_local(&store, &client, false, false).await.unwrap();
-
-        let reqs = server.received_requests().await.unwrap();
-        let body = String::from_utf8(reqs[0].body.clone()).unwrap();
-        assert!(
-            !body.contains("vector"),
-            "server without the capability must get a text-only push: {body}"
-        );
-    }
-
-    /// A note queued for push with NO `note_embeddings` row at all (never
-    /// embedded locally, or embedding failed) must fall back to a text-only
-    /// push for that row — not crash, and not send an empty/malformed
-    /// `vector` field — even though the server accepts pushed vectors.
-    #[tokio::test]
-    async fn push_local_falls_back_to_text_only_when_local_embedding_missing() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        // Deliberately no `insert_embedding` call — this note has never been
-        // embedded.
-        store
-            .add_note("decision", "Unembedded", "first", &[], &[], None, None)
-            .unwrap();
-        let rows = store.rows_for_sync(false).unwrap();
-        assert_eq!(rows.len(), 1);
-        let uuid = rows[0].uuid.clone();
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 0,
-                "results": [{"status": "created", "external_id": uuid, "id": "cloud-1"}]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        // accepts_pushed_vectors = true, but no local embedding exists.
-        let summary = push_local(&store, &client, false, true).await.unwrap();
-        assert_eq!(
-            (summary.attempted, summary.created, summary.failed),
-            (1, 1, 0)
-        );
-
-        let reqs = server.received_requests().await.unwrap();
-        let body = String::from_utf8(reqs[0].body.clone()).unwrap();
-        assert!(
-            !body.contains("vector"),
-            "a note with no local embedding must fall back to text-only, \
-             not error or send a malformed vector: {body}"
-        );
-    }
-
-    /// `note_embeddings` is a `vec0` virtual table with a `FLOAT[896]` column
-    /// (migration `004_memory.sql`) — sqlite-vec enforces that exact
-    /// dimension AT INSERT TIME, for every write path (there is only one:
-    /// `insert_embedding`). So a "leftover pre-896 768-dim row" — unlike the
-    /// code-chunk `embeddings` table, which DID have a legacy 768-dim era
-    /// with an explicit recreate-on-open migration in `db.rs` — can never
-    /// actually be written for memory notes: there was never a 768-dim
-    /// memory-embedding vintage to migrate from, and the store itself
-    /// refuses the write. Confirmed here rather than assumed, since it is
-    /// exactly the scenario `push_local`'s dimension guard names in its
-    /// comment.
-    #[tokio::test]
-    async fn insert_embedding_rejects_wrong_dimension_vector() {
-        use tempfile::TempDir;
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        let rows = store.rows_for_sync(false).unwrap();
-
-        let stale_768_blob = spelunk_core::embeddings::vec_to_blob(&vec![1.0f32; 768]);
-        let err = store
-            .insert_embedding(rows[0].local_id, &stale_768_blob)
-            .unwrap_err();
-        assert!(
-            err.to_string().contains("896") && err.to_string().contains("768"),
-            "the vec0 FLOAT[896] column must refuse a 768-dim insert outright \
-             (this is what makes a wrong-dimension row unreachable via any \
-             application write path): {err}"
-        );
-    }
-
-    /// Since a wrong-dimension row can never be *written* (previous test),
-    /// the only way `push_local`'s guard (`blob_to_vec` decode + a length
-    /// check against `EMBEDDING_DIM` — the same two building blocks
-    /// exercised inline at `sync.rs`'s vector-resolution site) could ever
-    /// see a wrong-length vector is an on-disk blob corrupted or torn
-    /// independently of any insert (disk fault, a crash mid-write). This
-    /// pins that composed guard logic directly: it must never panic on a
-    /// short/truncated/empty blob, and must always filter such a decode out
-    /// rather than accept a spurious length or garbage-padded 896 vector.
-    #[test]
-    fn dim_guard_logic_rejects_short_truncated_and_empty_blobs() {
-        use spelunk_core::embeddings::{EMBEDDING_DIM, blob_to_vec, vec_to_blob};
-
-        let guarded = |blob: &[u8]| -> Option<Vec<f32>> {
-            Some(blob_to_vec(blob)).filter(|v| v.len() == EMBEDDING_DIM)
-        };
-
-        // A stale 768-float blob (wrong dimension, but cleanly decodable).
-        let stale_768 = vec_to_blob(&vec![1.0f32; 768]);
-        assert!(
-            guarded(&stale_768).is_none(),
-            "a 768-float blob must be filtered out, not accepted"
-        );
-
-        // A torn write: a valid 896-dim blob with its last few bytes cut off.
-        let full = vec_to_blob(&vec![1.0f32; EMBEDDING_DIM]);
-        let truncated = &full[..full.len() - 10];
-        assert_ne!(
-            blob_to_vec(truncated).len(),
-            EMBEDDING_DIM,
-            "sanity: a truncated blob must decode to a non-896 length"
-        );
-        assert!(
-            guarded(truncated).is_none(),
-            "a truncated blob must never pass the dimension guard"
-        );
-
-        // A zero-length blob (e.g. corrupted read).
-        assert!(
-            guarded(&[]).is_none(),
-            "an empty blob must be filtered out, not treated as a valid vector"
-        );
-    }
-
-    // ── stamping must not trust a non-persisted status ─────────────────────
-    // A server can return a per-item `id` for an entry alongside a status
-    // that does not affirm durable persistence (aggregate `created: 0`).
-    // Stamping `remote_id` anyway would permanently exclude the row from
-    // `live` on every future push — the data could never be retried. Only
-    // `created`/`skipped` may stamp; a `failed` item carrying an `id` must be
-    // left unstamped.
-    #[tokio::test]
-    async fn push_local_does_not_stamp_remote_id_for_a_failed_status_item() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-
-        let rows = store.rows_for_sync(false).unwrap();
-        assert_eq!(rows.len(), 1);
-        let ext_a = rows[0].uuid.clone();
-        // The server hands back an `id` even though the entry was not
-        // durably persisted (`created: 0`, status "failed").
-        let cloud_a = "01890000-0000-7000-8000-0000000000b1";
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 0, "skipped": 0, "failed": 1,
-                "results": [
-                    {"status": "failed", "external_id": ext_a, "id": cloud_a},
-                ]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s1 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!((s1.attempted, s1.created, s1.skipped), (1, 0, 0));
-
-        // The row must NOT carry the id the server handed back — it stays
-        // retryable on the next push.
-        assert_eq!(store.note_id_for_remote_id(cloud_a).unwrap(), None);
-        let rows_after = store.rows_for_sync(false).unwrap();
-        assert_eq!(rows_after[0].remote_id, None);
-
-        // A re-push must still consider this row live (not already-synced).
-        let live_again: Vec<_> = rows_after
-            .iter()
-            .filter(|r| !r.archived && r.remote_id.is_none())
-            .collect();
-        assert_eq!(live_again.len(), 1, "unstamped row must remain retryable");
-    }
-
-    // ── counts must reconcile against results[], not the aggregate ints ────
-    // `BatchPushResult`'s `created`/`skipped` ints and its `results[]` array
-    // are independent wire fields — a server can send an aggregate
-    // `created: 0` for a batch whose `results[]` shows every entry durably
-    // persisted. A push summary built from the aggregate ints alone would
-    // read as "nothing landed"; it must instead read the true outcome off
-    // `results[].status`.
-    #[tokio::test]
-    async fn push_local_reconciles_counts_from_results_not_aggregate_ints() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        store
-            .add_note("note", "Two", "second", &[], &[], None, None)
-            .unwrap();
-
-        let rows = store.rows_for_sync(false).unwrap();
-        let (ext_a, ext_b) = (rows[0].uuid.clone(), rows[1].uuid.clone());
-        let cloud_a = "01890000-0000-7000-8000-0000000000c1";
-        let cloud_b = "01890000-0000-7000-8000-0000000000c2";
-
-        let server = MockServer::start().await;
-        // The aggregate ints understate what happened (`created: 0, skipped:
-        // 0`), but `results[]` shows both entries durably persisted.
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 0, "skipped": 0, "failed": 0,
-                "results": [
-                    {"status": "created", "external_id": ext_a, "id": cloud_a},
-                    {"status": "skipped", "external_id": ext_b, "id": cloud_b},
-                ]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s1 = push_local(&store, &client, false, false).await.unwrap();
-        // Reconciled from `results[]`, not the misleading aggregate zeros.
-        assert_eq!(
-            (s1.attempted, s1.created, s1.skipped, s1.failed),
-            (2, 1, 1, 0)
-        );
-        assert_eq!(
-            store.note_id_for_remote_id(cloud_a).unwrap(),
-            Some(rows[0].local_id)
-        );
-        assert_eq!(
-            store.note_id_for_remote_id(cloud_b).unwrap(),
-            Some(rows[1].local_id)
-        );
-    }
-
-    // ── a failed item must not mask other successes in the same batch ─────
-    // Mixed outcome: one entry lands, one doesn't. The failed item must stay
-    // unstamped (retryable) while the successful one is recorded — and the
-    // summary must show the real partial success, not a false "nothing
-    // happened" (which is what reading only the aggregate `created` count
-    // for a batch containing any failure could produce).
-    #[tokio::test]
-    async fn push_local_partial_failure_reports_the_real_successes() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        store
-            .add_note("note", "Two", "second", &[], &[], None, None)
-            .unwrap();
-
-        let rows = store.rows_for_sync(false).unwrap();
-        let (ext_a, ext_b) = (rows[0].uuid.clone(), rows[1].uuid.clone());
-        let cloud_a = "01890000-0000-7000-8000-0000000000d1";
-        let cloud_b = "01890000-0000-7000-8000-0000000000d2";
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 1,
-                "results": [
-                    {"status": "created", "external_id": ext_a, "id": cloud_a},
-                    {"status": "failed", "external_id": ext_b, "id": cloud_b},
-                ]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s1 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!(
-            (s1.attempted, s1.created, s1.skipped, s1.failed),
-            (2, 1, 0, 1),
-            "attempted must stay 2 (not read as nothing-to-push) and the \
-             genuine success must be visible alongside the failure"
-        );
-        // The successful row is stamped...
-        assert_eq!(
-            store.note_id_for_remote_id(cloud_a).unwrap(),
-            Some(rows[0].local_id)
-        );
-        // ...the failed one is not, and remains retryable.
-        assert_eq!(store.note_id_for_remote_id(cloud_b).unwrap(), None);
-        let rows_after = store.rows_for_sync(false).unwrap();
-        let live_again: Vec<_> = rows_after
-            .iter()
-            .filter(|r| !r.archived && r.remote_id.is_none())
-            .collect();
-        assert_eq!(live_again.len(), 1, "failed row must remain retryable");
-    }
-
-    // ── push_local's counting stays honest on a total-failure batch ────────
-    // `push_local` itself just reports honest counts (Bug 1/3's fix); it is
-    // the command layer (`memory_push` / `memory_sync`) that decides whether
-    // those counts mean "Done"/"Sync complete" or a hard failure, and that
-    // command-layer decision (the `bail!` that gives the CLI its non-zero
-    // exit) is covered end to end by the subprocess tests in
-    // `crates/spelunk-cli/tests/memory_push_sync_total_failure.rs`, not here.
-    // This test only pins `push_local`'s own return value for the all-failed
-    // shape those command-layer tests depend on.
-    #[tokio::test]
-    async fn push_local_total_failure_reports_zero_created_and_skipped() {
-        use tempfile::TempDir;
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        register_sqlite_vec();
-        let tmp = TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        store
-            .add_note("decision", "One", "first", &[], &[], None, None)
-            .unwrap();
-        store
-            .add_note("note", "Two", "second", &[], &[], None, None)
-            .unwrap();
-
-        let rows = store.rows_for_sync(false).unwrap();
-        let (ext_a, ext_b) = (rows[0].uuid.clone(), rows[1].uuid.clone());
-
-        let server = MockServer::start().await;
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 0, "skipped": 0, "failed": 2,
-                "results": [
-                    {"status": "failed", "external_id": ext_a, "id": serde_json::Value::Null},
-                    {"status": "failed", "external_id": ext_b, "id": serde_json::Value::Null},
-                ]
-            })))
-            .mount(&server)
-            .await;
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-
-        let s1 = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!(
-            (s1.attempted, s1.created, s1.skipped, s1.failed),
-            (2, 0, 0, 2),
-            "total failure: attempted > 0 but nothing durably landed"
-        );
-        // Neither row is stamped — both remain retryable.
-        let rows_after = store.rows_for_sync(false).unwrap();
-        assert!(rows_after.iter().all(|r| r.remote_id.is_none()));
-    }
-
-    // ── real-server regression: an established client must keep pulling ────
-    // A wiremock stand-in can't catch this class of bug: it lives in the real
-    // server's own handler/db pairing (a batch-push ack echoing the raw row
-    // id instead of the `sync_id` `/memory/since` cursors on), not in the wire
-    // shape a hand-typed mock response can get right by construction. Spins
-    // up the actual `spelunk-server` axum router, matching the pattern in
-    // `outbox.rs`'s `spawn_spelunk_server`.
-
-    /// Spin up a real `spelunk-server` axum router (the production router) on
-    /// an ephemeral loopback port, serving the team-hosting
-    /// `/v1/projects/*/memory*` routes this test's `CloudSyncClient`s talk to.
-    async fn spawn_spelunk_server() -> std::net::SocketAddr {
-        register_sqlite_vec();
-        let db_dir = tempfile::TempDir::new().unwrap();
-        let db =
-            spelunk_server::db::ServerDb::open(&db_dir.path().join("server.db"), 4, "test-model")
-                .unwrap();
-        let instance_id = db.get_or_create_instance_id().unwrap();
-        let state = spelunk_server::AppState {
-            db: std::sync::Arc::new(tokio::sync::Mutex::new(db)),
-            auth: std::sync::Arc::new(spelunk_server::auth::ApiKeyAuth::new(None)),
-            conflict_threshold: spelunk_server::default_conflict_threshold(),
-            embedder: spelunk_server::EmbedderSlot::disabled(),
-            embed_admission: spelunk_server::EmbedAdmission::new(
-                spelunk_server::EMBED_QUEUE_CAPACITY,
-                spelunk_server::EMBED_BUSY_RETRY_AFTER_SECS,
-            ),
-            llm: None,
-            max_tokens_ceiling: 8192,
-            rate_limiter: std::sync::Arc::new(spelunk_server::rate_limiter::RateLimiter::new(
-                1000, 60,
-            )),
-            instance_id,
-            started_by: None,
-            relay: spelunk_server::relay::RelayRegistry::new(),
-        };
-        let app = spelunk_server::router(state);
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let addr = listener.local_addr().unwrap();
-        tokio::spawn(async move {
-            let _ = axum::serve(listener, app).await;
-        });
-        addr
-    }
-
-    /// Reproduces the walk-the-store bug: a client that has already pushed
-    /// and synced once (an "established" client) never sees a teammate's
-    /// later entries on a subsequent sync, even though a fresh client would
-    /// pull the full set. Before the fix, client A's first push stamps its
-    /// own row's `remote_id` from the batch ack's `id` field, which the real
-    /// server (bug) fills with the raw autoincrement row id ("1", "2", ...)
-    /// instead of `sync_id`. That digit-string sorts lexically AFTER every
-    /// real UUIDv7 `sync_id` (which starts with a much smaller hex nibble for
-    /// any current timestamp), so `max_remote_id()`'s cursor becomes that row
-    /// id and `since_id=<cursor>` on the second pull matches nothing, even
-    /// though the server holds teammate B's newer entry.
-    #[tokio::test]
-    async fn established_client_pulls_teammates_entries_added_after_its_first_sync() {
-        register_sqlite_vec();
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        // Client A: first sync. Pushes its own entry, then pulls (nothing new
-        // yet): this is what "establishes" the client and stamps its remote_id.
-        let tmp_a = tempfile::TempDir::new().unwrap();
-        let store_a = MemoryStore::open(&tmp_a.path().join("memory.db")).unwrap();
-        store_a
-            .add_note(
-                "decision",
-                "A1",
-                "client A's own entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        let client_a = CloudSyncClient::new(&base_url, "proj", None, None).unwrap();
-
-        let push1 = push_local(&store_a, &client_a, false, false).await.unwrap();
-        assert_eq!(
-            push1.created, 1,
-            "client A's own entry must land on the server"
-        );
-        let pull1 = pull_and_apply(&store_a, &client_a).await.unwrap();
-        assert_eq!(pull1, 0, "nothing new on the server yet for the first pull");
-
-        // Teammate B: a second, independent client pushes a new entry to the
-        // same server/project.
-        let tmp_b = tempfile::TempDir::new().unwrap();
-        let store_b = MemoryStore::open(&tmp_b.path().join("memory.db")).unwrap();
-        store_b
-            .add_note("decision", "B1", "teammate B's entry", &[], &[], None, None)
-            .unwrap();
-        let client_b = CloudSyncClient::new(&base_url, "proj", None, None).unwrap();
-        let push_b = push_local(&store_b, &client_b, false, false).await.unwrap();
-        assert_eq!(
-            push_b.created, 1,
-            "teammate B's entry must land on the server"
-        );
-
-        // Client A syncs again: it is now an established client (already has a
-        // remote_id-stamped row), exactly the steady-state "sync to get
-        // teammates' latest" case the bug report describes.
-        let pull2 = pull_and_apply(&store_a, &client_a).await.unwrap();
-        assert_eq!(
-            pull2, 1,
-            "an established client must still pull entries a teammate pushed afterward"
-        );
-        let titles: Vec<String> = store_a
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.title)
-            .collect();
-        assert!(
-            titles.contains(&"B1".to_string()),
-            "client A must now have teammate B's entry locally: {titles:?}"
-        );
-    }
-
-    /// Three-way steady state, each established client syncing across
-    /// multiple rounds: rules out an off-by-one in cursor advancement that a
-    /// two-client, single-extra-round test (see above) could miss (e.g. a
-    /// cursor that only "catches up" once and then drifts on a later round).
-    ///
-    /// Client C joins second, via a pull-only first sync (see the doc
-    /// comment on `store_c`'s setup below for why: joining via push+pull in
-    /// one round is a separate, real ordering issue, not this story's bug).
-    /// After that, both A and C are established clients holding a cursor
-    /// derived purely from pulls/pushes of their own already-caught-up
-    /// state, and teammate B pushes twice, in two separate rounds. Each of A
-    /// and C must pick up exactly the right delta on each of their own
-    /// subsequent pulls: never 0 (the bug this story fixes), never a
-    /// duplicate re-application, and never the other established client's
-    /// own entries re-surfacing.
-    #[tokio::test]
-    async fn two_established_clients_each_pull_correctly_across_multiple_rounds() {
-        register_sqlite_vec();
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        // Client A establishes: push A1, pull (nothing yet).
-        let tmp_a = tempfile::TempDir::new().unwrap();
-        let store_a = MemoryStore::open(&tmp_a.path().join("memory.db")).unwrap();
-        store_a
-            .add_note("decision", "A1", "client A's entry", &[], &[], None, None)
-            .unwrap();
-        let client_a = CloudSyncClient::new(&base_url, "proj3", None, None).unwrap();
-        assert_eq!(
-            push_local(&store_a, &client_a, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-        assert_eq!(pull_and_apply(&store_a, &client_a).await.unwrap(), 0);
-
-        // Client C joins with no local content yet, so its first sync is a
-        // pull only (matching the walk-the-store "fresh client" case, which
-        // is known-good). This deliberately avoids a SEPARATE, real ordering
-        // issue that is out of scope for this story: `memory_sync` pushes
-        // before it pulls, so a client that pushes brand-new local content
-        // in the same round as older, not-yet-pulled remote content stamps
-        // its own freshly-minted (and therefore chronologically newest)
-        // sync_id as its cursor, permanently shadowing that older remote
-        // content from every future pull. Filed separately; see this task's
-        // board comment.
-        let tmp_c = tempfile::TempDir::new().unwrap();
-        let store_c = MemoryStore::open(&tmp_c.path().join("memory.db")).unwrap();
-        let client_c = CloudSyncClient::new(&base_url, "proj3", None, None).unwrap();
-        let pull_c1 = pull_and_apply(&store_c, &client_c).await.unwrap();
-        assert_eq!(pull_c1, 1, "client C must pull client A's A1 on establish");
-
-        // Now that C is caught up (nothing outstanding to miss), it can
-        // safely push its own new entry in the same round it pulls.
-        store_c
-            .add_note("decision", "C1", "client C's entry", &[], &[], None, None)
-            .unwrap();
-        assert_eq!(
-            push_local(&store_c, &client_c, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-        assert_eq!(
-            pull_and_apply(&store_c, &client_c).await.unwrap(),
-            0,
-            "nothing further for C to pull immediately after its own push"
-        );
-
-        // Teammate B pushes its first entry.
-        let tmp_b = tempfile::TempDir::new().unwrap();
-        let store_b = MemoryStore::open(&tmp_b.path().join("memory.db")).unwrap();
-        store_b
-            .add_note(
-                "decision",
-                "B1",
-                "teammate B's first entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        let client_b = CloudSyncClient::new(&base_url, "proj3", None, None).unwrap();
-        assert_eq!(
-            push_local(&store_b, &client_b, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-
-        // Round 2: both established clients must pick up exactly the new
-        // delta each is missing (A is missing C1 and B1; C is missing B1).
-        let pull_a_round2 = pull_and_apply(&store_a, &client_a).await.unwrap();
-        assert_eq!(
-            pull_a_round2, 2,
-            "client A must pull both C1 and B1 on its second sync"
-        );
-        let pull_c_round2 = pull_and_apply(&store_c, &client_c).await.unwrap();
-        assert_eq!(
-            pull_c_round2, 1,
-            "client C must pull only B1 (it already has A1 and its own C1)"
-        );
-
-        // Teammate B pushes a second entry. This is the off-by-one probe:
-        // a cursor that only advances correctly ONCE (round 2) but then
-        // sticks or drifts would surface here as 0 or a re-applied dup on
-        // this THIRD round for either established client.
-        store_b
-            .add_note(
-                "decision",
-                "B2",
-                "teammate B's second entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        assert_eq!(
-            push_local(&store_b, &client_b, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-
-        let pull_a_round3 = pull_and_apply(&store_a, &client_a).await.unwrap();
-        assert_eq!(
-            pull_a_round3, 1,
-            "client A's cursor must advance correctly again on a third round"
-        );
-        let pull_c_round3 = pull_and_apply(&store_c, &client_c).await.unwrap();
-        assert_eq!(
-            pull_c_round3, 1,
-            "client C's cursor must advance correctly again on a third round"
-        );
-
-        let titles_a: Vec<String> = store_a
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.title)
-            .collect();
-        assert!(
-            ["A1", "C1", "B1", "B2"]
-                .iter()
-                .all(|t| titles_a.contains(&t.to_string())),
-            "client A must end up with all four entries exactly once each: {titles_a:?}"
-        );
-        let titles_c: Vec<String> = store_c
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.title)
-            .collect();
-        assert!(
-            ["A1", "C1", "B1", "B2"]
-                .iter()
-                .all(|t| titles_c.contains(&t.to_string())),
-            "client C must end up with all four entries exactly once each: {titles_c:?}"
-        );
-    }
-
-    // ── sync_round: two-phase reconciliation ────────────────────────────────
-    // `sync_round` is `memory_sync`'s actual push+pull sequence, extracted so
-    // it can be driven directly against a real spawned server. `memory_sync`
-    // itself can't cheaply carry these scenarios: `capability::get_tier`
-    // caches its probe result in a per-process `OnceCell`, so several
-    // differently-configured in-process probes in one test binary would see
-    // stale tiers from whichever test's probe ran first.
-
-    /// Open a fresh local memory store in a new tempdir, returning both (the
-    /// tempdir must be kept alive by the caller for the store's lifetime).
-    fn fresh_store() -> (tempfile::TempDir, MemoryStore) {
-        register_sqlite_vec();
-        let tmp = tempfile::TempDir::new().unwrap();
-        let store = MemoryStore::open(&tmp.path().join("memory.db")).unwrap();
-        (tmp, store)
-    }
-
-    /// The primary repro, fixed: a client with local-only, never-pushed
-    /// content, running the actual `sync_round` sequence against a project
-    /// that already has a teammate's prior entry (pushed strictly before
-    /// this round begins),
-    /// ends the round with that teammate entry applied - not 0. This is
-    /// exactly the case the existing `two_established_clients_...` test
-    /// deliberately routes around (see its own comment) because, before this
-    /// fix, `memory_sync`'s push-then-pull order shadowed it permanently.
-    #[tokio::test]
-    async fn sync_round_pulls_teammates_prior_entry_on_a_first_round_with_local_content() {
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        // Teammate A establishes the project first, entirely before client
-        // C's own sync round begins.
-        let (_tmp_a, store_a) = fresh_store();
-        store_a
-            .add_note(
-                "decision",
-                "A1",
-                "teammate's prior entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        let client_a = CloudSyncClient::new(&base_url, "proj-primary", None, None).unwrap();
-        assert_eq!(
-            push_local(&store_a, &client_a, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-
-        // Client C has its own never-pushed local entry and has never synced.
-        let (_tmp_c, store_c) = fresh_store();
-        store_c
-            .add_note(
-                "decision",
-                "C1",
-                "client C's own new entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        let client_c = CloudSyncClient::new(&base_url, "proj-primary", None, None).unwrap();
-
-        let outcome = sync_round(&store_c, &client_c, false, false).await.unwrap();
-        assert_eq!(outcome.pushed.created, 1, "C's own entry must land");
-        assert_eq!(
-            outcome.pulled, 1,
-            "C must pull A's prior entry within this same round, not 0"
-        );
-        let titles: Vec<String> = store_c
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.title)
-            .collect();
-        assert!(titles.contains(&"A1".to_string()) && titles.contains(&"C1".to_string()));
-    }
-
-    /// Idempotence + no double-counting: running `sync_round` twice back to
-    /// back with nothing new to push or pull is a no-op both times, and the
-    /// round's own just-pushed row (harmlessly re-fetched by the second,
-    /// pre-round-cursor pull) is never counted twice or duplicated locally.
-    #[tokio::test]
-    async fn sync_round_twice_with_nothing_new_is_idempotent_and_never_double_counts() {
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        let (_tmp, store) = fresh_store();
-        store
-            .add_note("decision", "A1", "own entry", &[], &[], None, None)
-            .unwrap();
-        let client = CloudSyncClient::new(&base_url, "proj-idem", None, None).unwrap();
-
-        let r1 = sync_round(&store, &client, false, false).await.unwrap();
-        assert_eq!(r1.pushed.created, 1);
-        assert_eq!(
-            r1.pulled, 0,
-            "the second pull re-fetches this round's own just-pushed row via \
-             the pre-round cursor, but it must not be double-counted"
-        );
-        assert_eq!(store.count().unwrap(), 1, "no duplicate local row");
-
-        let r2 = sync_round(&store, &client, false, false).await.unwrap();
-        assert_eq!(
-            (r2.pushed.attempted, r2.pushed.already_synced, r2.pulled),
-            (0, 1, 0),
-            "a second round with nothing new must be a full no-op"
-        );
-        assert_eq!(store.count().unwrap(), 1);
-    }
-
-    /// The race window a plain reorder cannot close: a teammate's push
-    /// that lands on the server strictly between this round's own first pull
-    /// and its own push must still be picked up within this same round (via
-    /// the second pull, reusing the pre-round cursor) rather than being
-    /// permanently shadowed by the round's own push becoming the new
-    /// `MAX(remote_id)`.
-    ///
-    /// Real network concurrency can't be forced deterministically in a unit
-    /// test, so this composes `sync_round`'s exact same three calls
-    /// (`pull_and_apply_since` / `push_local` / `pull_and_apply_since`,
-    /// reusing one `pre_round_cursor`) with the teammate's push manually
-    /// interleaved at the precise point the race window occupies.
-    #[tokio::test]
-    async fn sync_round_catches_a_teammate_push_landing_between_its_own_pull_and_push() {
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        let (_tmp, store) = fresh_store();
-        store
-            .add_note("decision", "Client1", "own new entry", &[], &[], None, None)
-            .unwrap();
-        let client = CloudSyncClient::new(&base_url, "proj-race", None, None).unwrap();
-
-        // Step 1 of sync_round: capture the cursor, then pull. Nothing on the
-        // server yet.
-        let pre_round_cursor = store.max_remote_id().unwrap();
-        let pulled_first = pull_and_apply_since(&store, &client, pre_round_cursor.as_deref())
-            .await
-            .unwrap();
-        assert_eq!(pulled_first, 0);
-
-        // The race window: a teammate pushes here, strictly between this
-        // round's own pull and its own push.
-        let (_tmp_b, store_b) = fresh_store();
-        store_b
-            .add_note(
-                "decision",
-                "B1",
-                "teammate's race-window entry",
-                &[],
-                &[],
-                None,
-                None,
-            )
-            .unwrap();
-        let client_b = CloudSyncClient::new(&base_url, "proj-race", None, None).unwrap();
-        assert_eq!(
-            push_local(&store_b, &client_b, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-
-        // Step 2 of sync_round: this round's own push.
-        let pushed = push_local(&store, &client, false, false).await.unwrap();
-        assert_eq!(pushed.created, 1);
-
-        // Step 3 of sync_round: the second pull, reusing pre_round_cursor
-        // (NOT a freshly re-derived max_remote_id(), which would now include
-        // this round's own push and shadow B1 forever).
-        let pulled_second = pull_and_apply_since(&store, &client, pre_round_cursor.as_deref())
-            .await
-            .unwrap();
-        assert_eq!(
-            pulled_second, 1,
-            "the race-window teammate push must be caught by the second pull, \
-             not permanently lost"
-        );
-
-        let titles: Vec<String> = store
-            .rows_for_sync(false)
-            .unwrap()
-            .into_iter()
-            .map(|r| r.title)
-            .collect();
-        assert!(titles.contains(&"B1".to_string()));
-    }
-
-    /// `memory pull` (one-way, no push) is unaffected by the `sync_round`
-    /// two-phase reconciliation added for `sync`. It keeps
-    /// deriving a single cursor from the store itself via `pull_and_apply`,
-    /// unmodified.
-    #[tokio::test]
-    async fn pull_and_apply_one_way_pull_still_derives_its_own_single_cursor() {
-        let addr = spawn_spelunk_server().await;
-        let base_url = format!("http://{addr}");
-
-        let (_tmp_a, store_a) = fresh_store();
-        store_a
-            .add_note("decision", "A1", "first", &[], &[], None, None)
-            .unwrap();
-        let client_a = CloudSyncClient::new(&base_url, "proj-pull", None, None).unwrap();
-        assert_eq!(
-            push_local(&store_a, &client_a, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-        store_a
-            .add_note("decision", "A2", "second", &[], &[], None, None)
-            .unwrap();
-        assert_eq!(
-            push_local(&store_a, &client_a, false, false)
-                .await
-                .unwrap()
-                .created,
-            1
-        );
-
-        // A pull-only client with nothing local picks up both in one call.
-        let (_tmp_c, store_c) = fresh_store();
-        let client_c = CloudSyncClient::new(&base_url, "proj-pull", None, None).unwrap();
-        let pulled = pull_and_apply(&store_c, &client_c).await.unwrap();
-        assert_eq!(pulled, 2);
-
-        // A second, immediate pull is a no-op (cursor re-derived from what
-        // was just applied).
-        let pulled_again = pull_and_apply(&store_c, &client_c).await.unwrap();
-        assert_eq!(pulled_again, 0);
-    }
-
-    // ── pull pagination: exhaust every page, not just the first ─────────────
-    // Before this fix, `pull_and_apply_since` made exactly one request to
-    // `pull_since` and treated whatever it returned as the whole backlog. The
-    // server caps a single response at `CloudSyncClient::PULL_PAGE_LIMIT`
-    // entries, so a first sync into an established project silently applied
-    // only the first page and reported success. These tests drive the fix
-    // directly against a mock `/memory/since`, controlling exact page sizes
-    // (including ones at/above the real 500 clamp) without needing that many
-    // real rows in a live server.
-
-    /// Deterministic, lexically-increasing ids so `since_id` cursors compare
-    /// the same way real UUIDv7 cloud ids do.
-    fn page_ids(start: usize, count: usize) -> Vec<String> {
-        (start..start + count)
-            .map(|i| format!("01890000-0000-7000-8000-{i:012x}"))
-            .collect()
-    }
-
-    fn entries_json(ids: &[String]) -> serde_json::Value {
-        let entries: Vec<_> = ids
-            .iter()
-            .map(|id| {
-                serde_json::json!({
-                    "id": id,
-                    "kind": "note",
-                    "title": format!("T-{id}"),
-                    "body": "b",
-                    "created_at": "2026-06-19T01:00:00Z",
-                })
-            })
-            .collect();
-        serde_json::json!({ "entries": entries, "count": entries.len() })
-    }
-
-    const NIL_UUID: &str = "00000000-0000-0000-0000-000000000000";
-
-    /// Mounts one `/memory/since` mock per page, matched by the exact
-    /// `since_id` it must be requested with (the prior page's last id, or the
-    /// nil UUID for the very first request). Each mock must be hit exactly
-    /// `times` times.
-    async fn mount_pages_times(server: &wiremock::MockServer, pages: &[Vec<String>], times: u64) {
-        use wiremock::matchers::{method, path, query_param};
-        use wiremock::{Mock, ResponseTemplate};
-
-        let mut cursor = NIL_UUID.to_string();
-        for ids in pages {
-            Mock::given(method("GET"))
-                .and(path("/v1/projects/proj/memory/since"))
-                .and(query_param("since_id", cursor.clone()))
-                .respond_with(ResponseTemplate::new(200).set_body_json(entries_json(ids)))
-                .expect(times)
-                .mount(server)
-                .await;
-            if let Some(last) = ids.last() {
-                cursor = last.clone();
-            }
-        }
-    }
-
-    async fn mount_pages(server: &wiremock::MockServer, pages: &[Vec<String>]) {
-        mount_pages_times(server, pages, 1).await;
-    }
-
-    /// Item 1: a backlog smaller than one page is a single request, and every
-    /// entry lands.
-    #[tokio::test]
-    async fn pull_and_apply_since_single_page_matches_prior_behavior() {
-        let server = wiremock::MockServer::start().await;
-        let page = page_ids(0, 40);
-        mount_pages(&server, &[page]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 40);
-        assert_eq!(server.received_requests().await.unwrap().len(), 1);
-    }
-
-    /// Item 2: a backlog spanning exactly two pages (500 then 150, the real
-    /// server clamp) is fetched in two requests, the second cursor is the
-    /// first page's last id, and every entry is applied exactly once.
-    #[tokio::test]
-    async fn pull_and_apply_since_two_pages_advances_cursor_to_last_id_of_prior_page() {
-        let server = wiremock::MockServer::start().await;
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 150);
-        mount_pages(&server, &[page1.clone(), page2.clone()]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 650);
-        assert_eq!(store.count().unwrap(), 650, "no duplicates applied");
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    /// Item 3: three-plus pages (500 + 500 + 200) keep looping past two
-    /// iterations, not just handling the two-page case.
-    #[tokio::test]
-    async fn pull_and_apply_since_three_pages_loops_past_two_iterations() {
-        let server = wiremock::MockServer::start().await;
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 500);
-        let page3 = page_ids(1000, 200);
-        mount_pages(&server, &[page1, page2, page3]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 1200);
-        assert_eq!(server.received_requests().await.unwrap().len(), 3);
-    }
-
-    /// Item 4: a page landing exactly on the limit is always followed by
-    /// exactly one more request (never assumed to be the last page), and
-    /// that follow-up returning short ends the loop immediately (never a
-    /// third, unnecessary request).
-    #[tokio::test]
-    async fn pull_and_apply_since_full_page_triggers_exactly_one_more_request() {
-        let server = wiremock::MockServer::start().await;
-        let page1 = page_ids(0, 500);
-        let page2: Vec<String> = vec![];
-        mount_pages(&server, &[page1, page2]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 500);
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    /// Item 5: an already fully-synced project (empty first page) terminates
-    /// after that one request instead of looping forever.
-    #[tokio::test]
-    async fn pull_and_apply_since_empty_first_page_terminates_after_one_request() {
-        let server = wiremock::MockServer::start().await;
-        mount_pages(&server, &[vec![]]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 0);
-        assert_eq!(server.received_requests().await.unwrap().len(), 1);
-    }
-
-    /// Item 6: `sync_round`'s reported `pulled` count (what `spelunk sync`'s
-    /// completion message prints) is the TRUE total across every page, not
-    /// just the first. `memory_sync` itself can't be driven directly in this
-    /// binary (see `sync_round`'s own doc comment on the per-process tier
-    /// cache), so this asserts on the exact value the message interpolates.
-    #[tokio::test]
-    async fn sync_round_pulled_count_reflects_every_page_not_just_the_first() {
-        let server = wiremock::MockServer::start().await;
-        // sync_round's own first pull pass sees the whole multi-page backlog
-        // (nothing local to push, so the push call is empty and the second
-        // pull pass, reusing the same pre-round cursor, will be a no-op
-        // repeat of the same now-caught-up query).
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 300);
-        {
-            use wiremock::matchers::{method, path, query_param};
-            use wiremock::{Mock, ResponseTemplate};
-            // No `expect(1)`: the confirmation pull re-derives from the SAME
-            // pre-round cursor (nil UUID, since nothing local existed before
-            // this round), so it legitimately repeats this identical
-            // two-page sequence a second time.
-            Mock::given(method("GET"))
-                .and(path("/v1/projects/proj/memory/since"))
-                .and(query_param("since_id", NIL_UUID))
-                .respond_with(ResponseTemplate::new(200).set_body_json(entries_json(&page1)))
-                .mount(&server)
-                .await;
-            Mock::given(method("GET"))
-                .and(path("/v1/projects/proj/memory/since"))
-                .and(query_param("since_id", page1.last().unwrap().clone()))
-                .respond_with(ResponseTemplate::new(200).set_body_json(entries_json(&page2)))
-                .mount(&server)
-                .await;
-        }
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let outcome = sync_round(&store, &client, false, false).await.unwrap();
-
-        assert_eq!(
-            outcome.pulled, 800,
-            "both pull passes re-fetch the same 800-entry backlog off the \
-             unchanged pre-round cursor; apply_remote_note's dedupe means the \
-             SECOND pass applies 0 new rows, so pulled must be exactly the \
-             true total, not double-counted nor short of the second page"
-        );
-        assert_eq!(store.count().unwrap(), 800);
-    }
-
-    /// Item 7: the one-way `spelunk memory pull` entry point (`pull_and_apply`,
-    /// which derives its own cursor from the store rather than being handed
-    /// one) also paginates fully — proven through `pull_and_apply` directly,
-    /// not just through `sync_round`, since both merely wrap the same shared
-    /// `pull_and_apply_since`.
-    #[tokio::test]
-    async fn pull_and_apply_one_way_also_paginates_fully() {
-        let server = wiremock::MockServer::start().await;
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 120);
-        mount_pages(&server, &[page1, page2]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply(&store, &client).await.unwrap();
-
-        assert_eq!(applied, 620);
-        assert_eq!(server.received_requests().await.unwrap().len(), 2);
-    }
-
-    /// Item 8: `sync_round`'s SECOND (confirmation) pull pass paginates fully
-    /// on its own when it independently has more than one page of results
-    /// (e.g. a teammate pushed a large batch in the gap between this round's
-    /// first pull and its push) — not just the first pull pass.
-    #[tokio::test]
-    async fn sync_round_second_pull_pass_paginates_fully_on_its_own() {
-        let server = wiremock::MockServer::start().await;
-        {
-            use wiremock::matchers::{method, path, query_param};
-            use wiremock::{Mock, ResponseTemplate};
-            // First pull pass (pre-push): nothing yet.
-            Mock::given(method("GET"))
-                .and(path("/v1/projects/proj/memory/since"))
-                .and(query_param("since_id", NIL_UUID))
-                .respond_with(
-                    ResponseTemplate::new(200)
-                        .set_body_json(serde_json::json!({ "entries": [], "count": 0 })),
-                )
-                .up_to_n_times(1)
-                .mount(&server)
-                .await;
-            // This round's own push (empty local store: nothing to push).
-            Mock::given(method("POST"))
-                .and(path("/v1/projects/proj/memory/batch"))
-                .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                    "created": 0, "skipped": 0, "failed": 0, "results": []
-                })))
-                .mount(&server)
-                .await;
-        }
-        // Second (confirmation) pull pass reuses the SAME pre-round cursor
-        // (nil UUID, since nothing was applied by the first pass): a
-        // teammate landed a 650-entry batch in the gap, spanning two pages.
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 150);
-        mount_pages(&server, &[page1, page2]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let outcome = sync_round(&store, &client, false, false).await.unwrap();
-
-        assert_eq!(outcome.pushed.attempted, 0);
-        assert_eq!(
-            outcome.pulled, 650,
-            "the confirmation pull pass must exhaust its own pagination too"
-        );
-        assert_eq!(store.count().unwrap(), 650);
-    }
-
-    /// Item 9: re-running a pull after an interrupted/partial prior run
-    /// (some entries already applied locally from an earlier page) must not
-    /// double-count `applied` for entries seen again — regression-guards the
-    /// existing dedupe-by-`remote_id` specifically under the new loop, since
-    /// a naive re-implementation could re-tally a page it had already
-    /// applied once before.
-    #[tokio::test]
-    async fn pull_and_apply_since_rerun_after_partial_prior_run_does_not_double_count() {
-        let server = wiremock::MockServer::start().await;
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 100);
-        // Each page is legitimately re-requested once per run below.
-        mount_pages_times(&server, &[page1.clone(), page2.clone()], 2).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let first_run = pull_and_apply_since(&store, &client, None).await.unwrap();
-        assert_eq!(first_run, 600);
-
-        // Re-running from the same (still `None`, un-advanced) cursor
-        // re-fetches the identical two pages; every entry is already known
-        // by `remote_id`, so nothing should be counted or inserted twice.
-        let rerun = pull_and_apply_since(&store, &client, None).await.unwrap();
-        assert_eq!(rerun, 0, "already-applied entries must not be re-counted");
-        assert_eq!(store.count().unwrap(), 600, "and never re-inserted");
-    }
-
-    /// Item 10a: a project not yet created on the server (404 on the very
-    /// first page request) still applies 0 and returns success, not an
-    /// error, unchanged by the new loop.
-    #[tokio::test]
-    async fn pull_and_apply_since_404_on_first_page_is_still_zero_not_an_error() {
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-        let server = MockServer::start().await;
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .respond_with(ResponseTemplate::new(404))
-            .expect(1)
-            .mount(&server)
-            .await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 0);
-    }
-
-    /// Item 10b: a `None` cursor still starts the loop's first request from
-    /// the nil UUID (full catch-up), unchanged by the new loop wrapping the
-    /// cursor for its own subsequent iterations.
-    #[tokio::test]
-    async fn pull_and_apply_since_none_cursor_still_starts_at_nil_uuid() {
-        let server = wiremock::MockServer::start().await;
-        mount_pages(&server, &[page_ids(0, 5)]).await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 5);
-    }
-
-    /// If the SECOND pull (the confirmation pull after this round's own push)
-    /// fails, `sync_round` must not silently swallow that error — but it also
-    /// must not lose or misrepresent the push that already succeeded. The
-    /// push already durably landed server-side and already stamped this
-    /// round's row with its `remote_id` (inside `push_local`, which returns
-    /// before the second pull ever runs), so: (1) the error surfaces instead
-    /// of being dropped, (2) its message says the push already reached the
-    /// server rather than reading as "nothing happened", and (3) local state
-    /// is left exactly as `push_local` left it — no corruption, no
-    /// re-attempt needed, just a retryable pull on the next sync.
-    #[tokio::test]
-    async fn sync_round_second_pull_failure_surfaces_the_error_without_losing_the_push() {
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let (_tmp, store) = fresh_store();
-        store
-            .add_note("decision", "T1", "own new entry", &[], &[], None, None)
-            .unwrap();
-        let ext = store.rows_for_sync(false).unwrap()[0].uuid.clone();
-        let cloud_id = "01890000-0000-7000-8000-0000000000b1";
-
-        let server = MockServer::start().await;
-        // First pull (pre-push, empty server): 200 empty.
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .respond_with(
-                ResponseTemplate::new(200)
-                    .set_body_json(serde_json::json!({"entries": [], "count": 0})),
-            )
-            .up_to_n_times(1)
-            .mount(&server)
-            .await;
-        // The push itself succeeds and durably lands.
-        Mock::given(method("POST"))
-            .and(path("/v1/projects/proj/memory/batch"))
-            .respond_with(ResponseTemplate::new(207).set_body_json(serde_json::json!({
-                "created": 1, "skipped": 0, "failed": 0,
-                "results": [{"status": "created", "external_id": ext, "id": cloud_id}]
-            })))
-            .mount(&server)
-            .await;
-        // The confirmation (second) pull hits a transient server error.
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .respond_with(ResponseTemplate::new(500))
-            .mount(&server)
-            .await;
-
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let err = sync_round(&store, &client, false, false)
-            .await
-            .expect_err("a real second-pull error must not be swallowed as success");
-
-        let msg = format!("{err:#}");
-        assert!(
-            msg.contains("push already reached the server") && msg.contains("1 created"),
-            "error must say the push already succeeded, not read as a total \
-             failure: {msg}"
-        );
-
-        // The push's own effect is untouched by the later pull error: the row
-        // is stamped with its cloud id, exactly as `push_local` left it.
-        assert!(
-            store.note_id_for_remote_id(cloud_id).unwrap().is_some(),
-            "the already-succeeded push must not be undone or left unstamped \
-             just because the confirmation pull afterward failed"
-        );
-        assert_eq!(
-            store.count().unwrap(),
-            1,
-            "no duplicate/corrupted local row"
-        );
-    }
-
-    /// A network/server failure on a LATER page (not the first) must not
-    /// corrupt or silently drop the pages that already succeeded: the loop
-    /// applies each page as it arrives, so a first-page success is durably
-    /// in the local store even though the overall call returns `Err`. A
-    /// naive rewrite (e.g. buffering all pages before applying any) would
-    /// instead lose the first page's entries when a later page fails.
-    #[tokio::test]
-    async fn pull_and_apply_since_error_on_a_later_page_keeps_earlier_pages_applied_and_is_retryable()
-     {
-        use wiremock::matchers::{method, path, query_param};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        // Page 1 must be a FULL page (== PULL_PAGE_LIMIT): a short page 1
-        // would already be the natural last page and the loop would never
-        // even attempt a second request, defeating the point of this test.
-        let page1 = page_ids(0, 500);
-        let page2 = page_ids(500, 5);
-
-        // ── Run 1: page 1 succeeds and applies, page 2 fails (500). ──
-        let server1 = MockServer::start().await;
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .and(query_param("since_id", NIL_UUID))
-            .respond_with(ResponseTemplate::new(200).set_body_json(entries_json(&page1)))
-            .expect(1)
-            .mount(&server1)
-            .await;
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .and(query_param("since_id", page1.last().unwrap().clone()))
-            .respond_with(ResponseTemplate::new(500))
-            .expect(1)
-            .mount(&server1)
-            .await;
-
-        let (_tmp, store) = fresh_store();
-        let client1 = CloudSyncClient::new(&server1.uri(), "proj", None, None).unwrap();
-        pull_and_apply_since(&store, &client1, None)
-            .await
-            .expect_err("a later-page failure must surface as Err, not a silent partial success");
-
-        assert_eq!(
-            store.count().unwrap(),
-            500,
-            "page 1's entries must already be durably applied even though the \
-             overall call failed on page 2"
-        );
-        assert_eq!(
-            store.max_remote_id().unwrap().as_deref(),
-            Some(page1.last().unwrap().as_str()),
-            "the store-derived cursor reflects exactly the pages that landed, \
-             so a retry resumes from the right place"
-        );
-
-        // ── Run 2: a healthy server serves the remainder from the re-derived
-        // cursor; nothing from page 1 is re-applied or duplicated. ──
-        let server2 = MockServer::start().await;
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .and(query_param("since_id", page1.last().unwrap().clone()))
-            .respond_with(ResponseTemplate::new(200).set_body_json(entries_json(&page2)))
-            .expect(1)
-            .mount(&server2)
-            .await;
-        let client2 = CloudSyncClient::new(&server2.uri(), "proj", None, None).unwrap();
-
-        let cursor = store.max_remote_id().unwrap();
-        let applied = pull_and_apply_since(&store, &client2, cursor.as_deref())
-            .await
-            .unwrap();
-        assert_eq!(applied, 5, "the retry applies exactly the remainder");
-        assert_eq!(
-            store.count().unwrap(),
-            505,
-            "no duplicates from re-fetching across the two runs"
-        );
-    }
-
-    /// A page whose entries fail to deserialize (here: an entry missing the
-    /// required `id` field, the value pagination advances the cursor from)
-    /// must fail the WHOLE page atomically rather than partially applying
-    /// the entries that happened to parse fine. `SinceBody`/`RemoteEntry`
-    /// deserialize the full response body before `pull_and_apply_since` ever
-    /// sees a single entry, so this is really a regression guard: a future
-    /// change that streamed/parsed entries one at a time could silently
-    /// apply a prefix before hitting the bad entry.
-    #[tokio::test]
-    async fn pull_and_apply_since_malformed_entry_missing_id_fails_the_page_atomically() {
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        let malformed = serde_json::json!({
-            "entries": [
-                {
-                    "kind": "note",
-                    "title": "no id field at all",
-                    "body": "b",
-                    "created_at": "2026-06-19T01:00:00Z"
-                }
-            ],
-            "count": 1
-        });
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(malformed))
-            .expect(1)
-            .mount(&server)
-            .await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        pull_and_apply_since(&store, &client, None)
-            .await
-            .expect_err("a page that fails to parse must surface as Err");
-
-        assert_eq!(
-            store.count().unwrap(),
-            0,
-            "nothing from an unparseable page may be partially applied"
-        );
-    }
-
-    /// The server's `count` field is documented (see `pull_since`'s wire
-    /// comment) as redundant with `entries.len()`, never a "more remain"
-    /// signal — so the loop's termination must be driven by the actual
-    /// number of entries returned, not by trusting `count`. A server (or a
-    /// test double) that reports an inflated `count` alongside a genuinely
-    /// short `entries` array must still be treated as the last page.
-    #[tokio::test]
-    async fn pull_and_apply_since_terminates_on_actual_entries_len_not_a_lying_count_field() {
-        use wiremock::matchers::{method, path};
-        use wiremock::{Mock, MockServer, ResponseTemplate};
-
-        let server = MockServer::start().await;
-        let ids = page_ids(0, 50);
-        let mut body = entries_json(&ids);
-        // `entries.len() == 50` (well short of PULL_PAGE_LIMIT), but `count`
-        // falsely claims far more remain. If the loop ever keyed off `count`
-        // instead of the real page length, this would spin into a second,
-        // unexpected request against a server with nothing left to mount.
-        body["count"] = serde_json::json!(9999);
-        Mock::given(method("GET"))
-            .and(path("/v1/projects/proj/memory/since"))
-            .respond_with(ResponseTemplate::new(200).set_body_json(body))
-            .expect(1)
-            .mount(&server)
-            .await;
-
-        let (_tmp, store) = fresh_store();
-        let client = CloudSyncClient::new(&server.uri(), "proj", None, None).unwrap();
-        let applied = pull_and_apply_since(&store, &client, None).await.unwrap();
-
-        assert_eq!(applied, 50);
-        assert_eq!(
-            server.received_requests().await.unwrap().len(),
-            1,
-            "a lying count must not trigger a second request past the real \
-             short page"
-        );
-    }
-}
